Small fixes to the submission detection experiment
I will restart the experiment, swapping the order of the features.
The reason is that the first feature is having stranger results and I'd
like to see the impact of the other two features more clearly.

This CL fixes a small issue in kAutofillPreferSavedFormAsSubmittedForm
and performs additional cleanups.

Bug: 40281981
Change-Id: Ib8f2717673d44433d9de90c409cb93b5471a930b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6063488
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1393356}