Don't forward useless frame-detached submission signals to AutofillAgent
This source is having huge volume in metrics [1], most of them trivially
not being useful in submissions, judging from how many are accepted by
AutofillAgent.

This CL modifies the logic so that the signals are only forwarded when
FormTracker is tracking a form of some sort. This will also reduce the
noise in the added metric of the previous CL.

Note that this CL is a no-op technically since signals that used to pass
before and won't anymore would anyway be discarded by AutofillAgent
since there wouldn't be any `provisionally_saved_form()` to return as
submitted form.

[1] https://uma.googleplex.com/p/chrome/timeline_v2?sid=7805e6638ed24dfc765979e35700e2e4

Bug: 40281981
Change-Id: Id42f72986158d7bc045496b6b1b9b63aeea96065
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6074913
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1393358}