Dangling pointer in UpdateRecentVisitsFromHistoryDBTask
UpdateRecentVisitsFromHistoryDBTask represents a DB task that is
executed on a background thread. It holds a raw_ptr<> to
URLIndexPrivateData, which is accessed in DoneRunOnMainThread() after
the task has finished. The pointer becomes dangling if the
URLIndexPrivateData dies before the task completes.

This CL resolves this by replacing the raw_ptr<> with a
scoped_refptr<>. This works, since URLIndexPrivateData already is a
base::RefCountedThreadSafe<>.

[1] https://source.chromium.org/chromium/chromium/src/+/main:components/omnibox/browser/url_index_private_data.cc;l=350;drc=1651676a30cd7abcd177975f7cd0e37bd945f663

Fixed: 41485242
Change-Id: Ifab8c41614790ff9186cb9b5787cd01e62db2a90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6054360
Reviewed-by: Jennifer Serrano <jennserrano@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1393551}