[iOS][QD] Attempt to deflake testInactiveTabsForDeletion
In testInactiveTabsForDeletion, the deletion was hanging causing the
test to fail. In this CL, we attempt to fix it by scoping the deletion
to only the tabs, has DBD is known to sometimes hang in egTests if
history is deleted.

This CL also adds a few checks to the QDMediator to ignore some UI
updates if a deletion is ongoing.

Fixed: 374048360
Change-Id: I068a3c42b9ddad38199eabbd7e9849a8383cf417
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6074550
Commit-Queue: Filipa Senra <fsenra@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393560}