[Tab Switcher] Invalidate favicon when URL is updated
For tab group sync the way the tab is navigated bypasses the normal
navigation method that resets the favicon. Normally this is fine since
the onFaviconUpdated is called. However, if the new page doesn't have a
favicon the favicon is not reset.

To fix this do what is done for thumbnails and reset the favicon on
URL change as well. If the URL has the same favicon no update will
happen. However, if the favicon differs this will ensure the correct
favicon is shown (or a globe placeholder). This isn't particularly
expensive since URLs rarely change while the tab switcher is open.

It does matter more for the bottom tab strip. However, this is the
correct behavior that should have been happening for the bottom tab
strip in the first place.

Note there are known bugs that still return the "wrong" favicon, but
this is happening in native and is unrelated to this change.

Fixed: 382520662
Change-Id: I86c50f7ec71bbe72db37a65919548900cc3eec2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6077546
Reviewed-by: Sky Malice <skym@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393687}