[Tab Switcher] Support TabCardLabelData in LIST mode
Add low-end support for TabCardLabelData by showing the label before
the list of URLs for a group in the list card items.

This largely reuses the code from GRID mode and just binds the view
differently as we have less control over the view as it is a modified
version of the modern list item / title and description layouts.

Fixed: 362306803
Change-Id: Ica6174c52b3bb90a46c64edcfeb5e7d2107c2e45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6077588
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1393707}