[FedCM] Reset verifying_focus_view_ when needed
In RemoveNonHeaderChildViewsAndUpdateHeaderIfNeeded(), we remove all
views that are not in the header. This means all pointers to views that
are not in the header need to be reset. The pointers are reorganized to
account for this distinction so it is harder to make the mistake in the
future. The error test is also augmented to include showing account UI,
which would surface this bug under ASAN bots.

Fixed: 382399969
Change-Id: I49c123e0b0f87f47715e14f7eb0b67b66eb9657e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6078188
Reviewed-by: Zachary Tan <tanzachary@chromium.org>
Commit-Queue: Nicolás Peña <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393744}