Prepare to binary-launch HistoryEmbeddings/Answers features
This CL prepares to default-enable HistoryEmbeddings and
HistoryEmbeddingsAnswers features for desktop platforms in US only.
Refactoring was necessary for the runtime logic used to launch these
features client-side on desktop platforms in the US.

Feature parameters are set to the launched Finch configuration values.
HistorySearchSettingsVisibility is not affected by this change.

Bug: 381108241
Change-Id: I6efd55508318edab47b4b02d5132ac930c8aa521
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6071094
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393792}