[A11y] Remove inaccuracies in name updates
This function ensures that if an object is labelled or described
by other objects, that if those other objects changed, that
the object with the dependent name is reserialized with the new
name, by marking it dirty. It walks up the parent chain from a
changed node (e.g. text node and looks for a focusable object that is
named by its descendants, or a label/description relation pointing to
the ancestor.

In the past this was the cause of performance issues, because we
did this for every changed node. However, now we only do it if the
changed node is part of a label or description, which is a cached
bit on each object. We return early if not.

Because of the early return, we can remove the innaccurate
optimizations:
- Limiting the # of ancestor that we walked (I found an example
on Twitter/X where the ancestor was higher than the max)
- If we found a dependent label/description, we stopped. However,
there could be more than one label or description that uses
the changed node.

Removing the inaccurateoptimizations do not make a difference in
the perf tests added with them (many-text-changes-*.html).

The benefit of doing this is that we should be able to ensure that
the browser-side a11y tree is always the same as the renderer side
tree.

Also cleaned up the code and addressed a todo.

Bug: none
Change-Id: I702f608eec5e920c8356dffbbe99d6d82f2c099c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6073166
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Jocelyn Tran <jocelyntran@google.com>
Cr-Commit-Position: refs/heads/main@{#1393845}