Add hash of url to keywords table
In order to detect corruption of this table, a new 'url_hash'
column is added to the 'keywords' table in 'Web Data'.

This contains an encrypted hash of 'url' and 'id'.

Records where this hash does not match are dropped when the
table is loaded.

The 'Web Data' database version is updated to 137 to reflect this
and tests are added to verify the new behavior.

Some minor refactoring is done to tidy some of the interfaces to
use modern semantics e.g. returning std::optional.

BUG=376303929

Change-Id: I374758e0e999c8548060b8735dac348786ad510e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6040381
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393869}