FSO: Hold FSAWatcherManager as a raw_ref in QuotaManager
We now pass a reference to the FSAWatcherManager when constructing
FileSystemAccessObserverQuotaManager. This is stored as a raw_ref in the
FileSystemAccessObserverQuotaManager. This works because the
FSAWatcherManager holds ObservationGroups. Each ObservationGroup has a
scoped_refptr to the QuotaManager for its StorageKey. When the last
ObservationGroup with a reference to a QuotaManager is removed, the
QuotaManager's ref count drops 0 and it is destructed. As such, the
FSAWatcherManager outlives the FileSystemAccessObserverQuotaManager.

This change also includes a testcase to ensure that the QuotaManager
removes a raw_ref to itself in the FSAWatcherManager when it is
destructed.

Bug: 338457523
Change-Id: Ib6e9e2721326523b8e2c438428359f4287be0b1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6075094
Commit-Queue: Rahul Singh <rahsin@microsoft.com>
Reviewed-by: Daseul Lee <dslee@chromium.org>
Reviewed-by: Nathan Memmott <memmott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393903}