[remoting] Support WirePlumber >= 0.5
The config template used by the Linux remoting host does not work with
newer versions of WirePlumber. Instead, this CL causes the host to run
WirePlumber with the default system config, but with a customized
profile.

The Debian package installs a new WirePlumber config "fragment" which
adds a custom "chrome-remote-desktop" profile. This will only be loaded
by newer WirePlumber versions which support config fragments.

Also, the environment variable $PIPEWIRE_REMOTE is added to the
session-manager child process, so that the virtual Audio/Sink device
is used by the session-manager and audio clients.

Bug: 372866283
Change-Id: I7e001bb5bee5f10265daa68caaf9e6bfeb639e57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6076056
Reviewed-by: Erik Jensen <rkjnsn@chromium.org>
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393978}