Allow WebXR layers to use any valid format.
Previously the use of CopyTextureToTexture internally meant that the
WebGPU layer format needed to match the backend swapchain exactly. This
change allows a different (less optimal) copy path if the format doesn't
match, meaning that we can allow either RGBA or BGRA formats.

Additionally, the preferred layer format now always matches the
preferred canvas format. This does mean that the "preferred" path isn't
currently the optimal one on Windows, but the intent is to fix that in
a future patch.

Finally, validation has been added to ensure that only the accepted formats are passed when creating the layer.

Bug: 359418629
Change-Id: I89ead830c7bf28c14867b14c3a8fd5dd9c1e0df7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6081975
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393988}