[lensoverlay] Migrate away from RunUntilIdle
RunUntilIdle is bad practice, leading to a presubmit warning anytime it
is used. This CL moves the query controller unittests away from this
bad practice, primarily relying on the TestFuture::Wait() instead.

The only exception is the suggest signals, which get called many times,
and therefore we cannot use a test future. Additionally, the callback
that runs does not wake up the RunUntil() run loop, so I had to use a
custom run loop and exit it once the suggest signals are ready.

Fixed: 374371322
Change-Id: I5f97e42ea7b6078404a6f4011315fd24b8f14436
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6075708
Commit-Queue: Duncan Mercer <mercerd@google.com>
Reviewed-by: Juan Mojica <juanmojica@google.com>
Cr-Commit-Position: refs/heads/main@{#1394020}