[anchor] Add IsAnchor(), IsExplicitAnchor(), fix HasAnchorQueryToPropagate().
In order for a fragment to be an anchor, it needs to be a CSS box (i.e.
not a line / fragmentainer - those share computed style with their
parent CSS box). There was already a (recently added) check for that in
FragmentBuilder::PropagateChildAnchors(), but other parts of the code,
i.e. StitchedAnchorQueries::AddOutOfFlowChild(), didn't properly check
for this. Now, that wasn't strictly necessary in that part of the code,
since we already check if the fragment is out-of-flow positioned, which
means that it cannot also be a fragmentainer. But I plan to use
HasAnchorQueryToPropagate() where this does matter, and I also intend
add code that checks if a fragment is an anchor, and rather than adding
more IsCSSBox() checks, let's do this.

No behavior changes intended.

Change-Id: I6d3d829eb37ebc93815ef74a0f1e926eac1729d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6080382
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394107}