[Tab Group Sync] Update tab attributions locally
This CL also introduces more tests to increase the coverage.

Shared attributions are used for local store only and they are not sent
further to the server, but the server may overwrite local attributions.

Note that as opposed to saved tab group attributions, shared
attributions are not automatically updated for the group on tab changes.

This CL also removes the test in SavedTabGroupBar related to sharing a
group because the tests use a real TabGroupSyncService and there is no
(simple) way to provide account ID from sync.

Bug: 352214654
Change-Id: Id8fe6c404a4bc418c3ae12923c486d33cd70988b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6069754
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Reviewed-by: David Pennington <dpenning@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394130}