Code Health: remove base::Feature "BlockPromptsIfIgnoredOften"
This feature is enabled by default for a very long time. It looks like it's safe to clean it up.

This CL also changes the signature of IsUnderEmbargo. That function is only called with flags that need to be deleted because they are enabled by default for a very long time.

Bug: 356236983
Change-Id: Ie919203db886d6c95b0bae6463adcd3b056084c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6071142
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Elias Klim <elklm@chromium.org>
Commit-Queue: Alexandra Reimers <areimers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394152}