Make the SelectorFilter pushing less lazy.
We had an elaborate scheme for pushing elements onto the SelectorFilter
only when needed, because style sharing could cause situations where
large trees would be processed but never actually needed style (and
thus, populating SelectorFilter would be a waste). This caused us to
have a tricky system of RAII objects linking to each others through
a global object and linked lists.

However, with style sharing long gone, the only case where we need
to be lazy is when we have an element with only text-node children,
which we can do more explicitly.

Style perftest is neutral or slightly positive.

Change-Id: I4b1167518d61f9b5653be315f6bbc06138f6e56c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6079413
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394202}