Fix a crash in CSS Custom Property application
The root highlight pseudo styles do not get recalculated when
a CSS custom property is registered, causing crashes when that
property is used in the pseudo class. Fix it by checking for a
change in InitialData in Element::CalculateHighlightRecalc.

Add a test for this case.

Bug: 381362600
Change-Id: Iabeaf88bedecfb0ee039b2e98151750844c7ab7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6077895
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394243}