[Settings/Autofill] Add Loyalty Cards count to Your Saved Info page
This change introduces a counter for Loyalty Cards from Google Wallet
to be displayed on the "Your saved info" settings page.

The count is fetched through the SavedInfoHandler which now observes
the ValuablesDataManager.
The count is updated in real-time to reflect changes from other tabs or sync.

This is the next step of Home of Transactions project.
This change only introduces a logic for reading the correct number
and makes the counter available on 'Your saved info' page as property.
It will be used and displayed on settings page as a follow-up.

Bug: 448351904
Change-Id: I3adc9455a3999e861be904589e1f25f29e885c75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7038035
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Stanislav Mikheyev <mikhe@google.com>
Commit-Queue: Ireneusz Szulc <ireneuszszulc@google.com>
Cr-Commit-Position: refs/heads/main@{#1530850}