Dictation: Remove mv2 tests
Since accessibility common has successfully launched in mv3, we are
removing the mv2 tests. This will help us remove duplicate tests and
unhook ourselves from the mv2 infrastructure, helping to unblock the
extensions team.

This change is scoped just to dictation; follow up changes will
remove mv2 tests for the other features in accessibility common.

Note that we will keep the mv2 implementation around for a
couple of milestones just in case we need to rollback in the case of
a severe regression.

Bug: 402207274
Change-Id: I5d724137040cca3f35791a94add1e4bb138cc613
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7031922
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530890}