[graphite] Add plumbing for IOSurface backpressure fences
This adds a gfx::MTLSharedEventFence type that wraps a MTLSharedEvent
and its associated value for waiting. This will be plumbed from the
OverlayRepresentation to the ImageTransportSurfaceOverlayMac instead of
via the GLContext as is done today. This CL only introduces the plumbing
and doesn't have any functional change.

Bug: 391990607
Change-Id: I23c52bb001cb2198da9098fa31a387d26a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7045912
Auto-Submit: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530979}