Remove chrome/browser/*browsertest* paths from requiring mutter CQ bot
The "linux-wayland-mutter-rel" CQ bot is often overloaded these days,
eg this build has 1+ hr pending and 1+ hr execution time:
https://ci.chromium.org/ui/p/chromium/builders/try/linux-wayland-mutter-rel/17135/infra

This is likely partly due to an under-provisioned test pool of
ubuntu-24.04 bots. When we replaced the linux-wayland-weston-rel bot on
the CQ with linux-wayland-mutter-rel, we provisioned an equivalent
amount of necessary test bots. But for whatever reasons, they don't
seem to be sufficient to avoid such pending times.

We can provision further ubuntu-24.04 machines for the mutter CQ bot,
but that will start dipping into our yearly organic machine quota. So
would be nice to avoid that.

So instead this reduces the amount of filepaths required for the
mutter CQ bot by just removing "chrome/browser/*browsertest*".

As-is, the current filepath list matches about ~50 CLs per day:
https://screenshot.googleplex.com/9KnwixztYNrPq9D

By removing just "chrome/browser/*browsertest*", that goes down to
~25 per day:
https://screenshot.googleplex.com/7VojrQ7rDJRKQC2

So halves the machine load by removing a single filepath. Hopefully
the machine savings outweighs the cost of potentially letting CLs that
only touch "chrome/browser/*browsertest*" in that regress the
mutter bot?

Bug: None
Change-Id: I2bf6ed27a3473295dc38f18974e449bd17078b59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7046968
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Kramer Ge <fangzhoug@chromium.org>
Reviewed-by: Orko Garai <orko@igalia.com>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530983}