[Theme] Change icon and text color with white background.
In this CL, we fix the icon and text colors issue on fake search box
and composeplate view on NTP when a white background is used in dark
theme:
- @color/default_icon_color_dark is used to tint all related
  icons buttons;
- New TextAppearances with @color/default_text_color_dark are used.
This allows all text and icons are used black color in both light/dark
modes when a white background is used for these two views.
Screenshots: http://shortn/_QBX1ccDK4t.

Bug: 423579377
Change-Id: I1d2d057d867fb0a12a49c6510af2a0435c92cad3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7045818
Commit-Queue: Xi Han <hanxi@chromium.org>
Reviewed-by: Xinyi Ji <xinyiji@chromium.org>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531035}