[PWA] Selectively re-enable sync integration tests on Linux DBG builds
Tests were flaky on the Linux (DBG) builds because it would take longer.
crrev.com/c/7049747 recently went in that increased the test timeout
limit of that bot, so selectively reenable all tests that have been
recently disabled to see if the flakiness goes down. This is a
speculative fix.

Bug: 452016158
Change-Id: I2124d73d8a0d2aebbcf1eaee06e9924c6b2ce9fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7049385
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531058}