Allow plugin to control unsafe-buffer-usage-in-unique-ptr-array diag.
This new diagnostic of the unsafe buffers family is present in recent
versions of clang. Our code is not compliant, and under some build
configurations of libc++, this is even a false positive.

Intercept the diagnostic and either ignore it, or promote it to
a warning, based upon a `unique_ptr` entry in the unsafe buffers
paths file (default is ignore).

-- Update .md file to be correct after a roll (currently the new
   directive is silently ignored, so should be ok).

Bug: 452084023
Change-Id: Ib87746e6a9f792eb5ec24abf0c3e433a23c6ecc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7048761
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531072}