Add a hit test readiness check for prerendering
This change introduces a callback mechanism to `ViewAndroid` to allow
`RenderWidgetHostViewAndroid` to determine if a view is ready for hit
testing.

A crash was observed in `RenderWidgetHostViewBase::ProcessTouchEvent`
when touch events were processed for a page that was in the process of
prerendering.

To fix this, a `HitTestCallback` is added to `ViewAndroid`.
`RenderWidgetHostViewAndroid` sets this callback to a function that
checks if the frame tree is currently prerendering. `ViewAndroid` then
uses this callback to determine if it should proceed with hit testing.
This prevents the crash by deferring touch event processing until the
page has finished prerendering.

Bug: 442832509
Change-Id: Ibf793dfa53ae38e198fd16e6ce24b1643d72a844
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7040604
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Keita Suzuki <suzukikeita@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531139}