net/disk_cache: Default DiskCacheBackendExperiment to the SQL backend
An upcoming field trial will enable the DiskCacheBackendExperiment
feature on WebView. However, with the current implementation, it is
impossible to manually select the SQL backend from WebView's flag page.

When the DiskCacheBackendExperiment flag is set to "Default", the
backend is chosen based on the field trial config. When set to "Enabled"
or "Disabled", it falls back to the simple backend because the default
value of the "backend" parameter is "kDefault".

This change modifies the default value of the "backend" parameter to
"kSql". This allows developers to explicitly enable the SQL backend by
setting the flag to "Enabled", which will be necessary for testing and
debugging the feature for the upcoming experiment.

Bug: 443174708
Change-Id: I729d75eaddf6a7c5d71042d3ac3d6ec591b5f606
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7046079
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531208}