extensions: Extract EAVC's stateful popup logic to platform delegate
This is a part of the patch series to move stateful UI logic from
ExtensionActionViewController to ExtensionActionPlatformDelegate.

This patch moves the implementation of stateful popup logic from
ExtensionActionViewController to ExtensionActionPlatformDelegateViews,
namely these functions:

  - IsShowingPopup()
  - HidePopup()
  - GetNativePopupView()
  - TriggerPopup()
  - ShowPopup()
  - OnPopupClosed()

Note that we keep the public API of ExtensionActionViewController
identical, though these functions now simply delegate to
ExtensionActionPlatformDelegate.

Importantly, most of ExtensionActionViewController::ExecuteUserAction()
is untouched so that its logic can be shared with non-Views-based
platforms.

This is a pure refactor. No functional changes are expected.

Bug: 448199168
Change-Id: I720684603836a7e013106ff2720517fc961e780a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7017009
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531231}