Update guidance on MiraclePtr usage
The style guide appears to be a bit stale, implying that `raw_ptr` and
`raw_ref` should not be used in renderer code. This has not been the
case for a while (though exceptions still apply).

This CL

*   updates the C++ style guide to capture the memory team's current
    stance,

*   Adds some examples for quick reference, and

*   trivially reflows some of the prose into bullet points.

As this guide is owned by cxx@, a discussion thread was posted at:
https://groups.google.com/a/chromium.org/g/cxx/c/8ZTKuZgoYNg/m/zkCUhO1TBwAJ

Bug: 40064499
Change-Id: Ieafbb16ead4d07371b116cb614f1a8b706c659d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7030781
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531302}