[mtb-cct] Record metric for sharing action origin
Add an intent extra in the sharing Intent indicating the origin
of the action. BrApp records the histogram so that we know
where in Chrome the action was originated. For now the main
interest is CCT custom button origin (11), and not all share
Intent objects are equipped with the new origin field.

Bug: 452687931
Change-Id: If0ef9a187cf4aa654ccdcfde4b201f95e5dd9a6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7050446
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531374}