webrtc-internals: add RTCStats dump format
This CL adds an alternative format for capturing the information gathered by webrtc-internals. The format is described at
  https://github.com/rtcstats/rtcstats/tree/main/packages/rtcstats-server#rtcstats-dump-file-format
In a nutshell it consists of
- a magic signature "RTCStatsDump"
- a line with a JSON object containing metadata
- a series of JSON arrays containing [event name, connection id, value, extra data..., timestamp delta] separated by newlines.

This format can (like the JSON one) be imported on
  https://rtcstats.github.io/rtcstats/dump-importer/
which supports both formats.

Also unify timestamp handling and add the page url to the getUserMedia logging.

Bug: chromium:450197205
Change-Id: Iaeb98db721dc400a60b1eef6bf70d1cf4d75ac1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7002775
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Commit-Queue: Philipp Hancke <philipp.hancke@googlemail.com>
Cr-Commit-Position: refs/heads/main@{#1531407}