[Task Icon] Use first available instance in GlicActorTaskIconManager
Change the GlicInstance check to use the first available instance for
the TaskIcon updates. Currently this check relies on the last actuated
tab, which isn't accessible after a task finishes. This is an
intermediate fix until the GlicInstance is available through the
ActorTask and will not work with multiple tasks. The redesigned task
nudge will not need a GlicInstance check and thus further work to fix
this in the existing Task Icon will not be used going forward.

Before:
https://screencast.googleplex.com/cast/NTc5MDM5NDg4ODQ4NjkxMnxiZDYzYmZkNy05Zg
After:
https://screencast.googleplex.com/cast/NTU4MTExOTg1NTUyNTg4OHw3NzU0ODk1OC0zNA

Bug: 452012830
Change-Id: Ic81a55fdd20265a1c31e4214b6d347f49d180ebb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7048044
Reviewed-by: Ian Wells <iwells@chromium.org>
Commit-Queue: Maggie Jennings <mjenn@google.com>
Reviewed-by: Alison Gale <agale@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531505}