[CSS Modules] Create Import Map entry for Declarative CSS Modules
This change adds the `specifier` attribute and creates an import map
based on the attribute value and the contents of the associated <style>
tag.

This behavior is currently gated behind the existing `style-src` CSP, as
regular style tags obey today, and is subject to change based on
feedback.

A few basic WPT's were added to test functionality (and CSP behavior),
as well as a unit test verifying the internal import map structure. More
WPT's will come in a follow-up.

Bug: 448174611
Change-Id: I68fa88d900a72be2e004249687912098c7a97812
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7038416
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1531546}