Force mono capture of mono output devices
This is a workaround for a bug where some mono streams are not upmixed
to stereo in case of a mono device loopback. With the workaround enabled
we always capture mono devices in mono and upmix to stereo in
CatapAudioInputStreamSource.

The workaround can be disabled by disabling the feature
kMacCatapForceMonoCapture.

Change-Id: I21617e4c9bae4aacab2533b6f49d48d244dd12ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037537
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Johannes Kron <kron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531550}