[ProfilePicker] Simplify startup logic for multiple profiles
Remove extra condition on users with multiple profiles that have
never seen the Profile Picker, to directly open a browser. They now
see the Profile Picker.

Details for the rational for the fix in the linked bug.

ProfilePicker::Shown() static function is still used to determine
which message to show on the Profile Picker (different message
on the first run of the Picker).
Added a pixel test for the scenario of second picker run (all existing
test assumes first picker run).

Fixed: 452609632
Change-Id: I7eea5bad9b500d501dfcdcb0fa2cee409f2d549e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7046483
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1531554}