legion: Add Crypter for post-handshake encryption
This change introduces a `Crypter` class to handle encryption and
decryption of messages after a secure channel has been established. It
uses AES-256-GCM for authenticated encryption. Messages are padded to a
multiple of 32 bytes, with the last byte of the padded message
indicating the number of zero-padding bytes. Nonces are derived from
incrementing sequence numbers for each encryption and decryption
operation.

Bug: b:444610593
Change-Id: Ia3f0dbb566a0d691dc041863ef96d251b039f63c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7043847
Reviewed-by: Mohamad Arab <arabm@google.com>
Commit-Queue: Nikhil Jakhar <nikhiljakhar@google.com>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531562}