Fix crash when closing side panel
The side panel was crashing after the refactor because after the close
button is clicked, it triggers the header to close the side panel.
This then leads to the controller to be destroyed and causes a seg
fault when the controller later tries to close the pin promo. This
CL fixes this bug by removing the explicit call to MaybeEndPinPromo()
after closing the side panel because the anchor to the promo has
been deleted so the promo also dismisses.

Fixed: 452911460
Change-Id: I62faaa3fe09eb8ef124406df59243c605c76d2f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7056147
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Commit-Queue: Steven Luong <stluong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531584}