[iOS]Uses check in bookmarks folder coordinator
In crbug.com/450910124 we see the mediator was not properly
disconnected. If this was a check, we would have probably seen the issue
is with the coordinator. This CL add checks for all instance variables,
for the sake of safety.

Bug: 450910124
Change-Id: I7bcca4787731580cbb28f5643a6d2dfe38c4071e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7047046
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: Huiting Yu <huitingyu@google.com>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531585}