[BNPL][Clank] Handle user clicking BNPL chip after AE completes
This CL handles the user flow for clicking the BNPL chip after the
transaction amount has been extracted.
When a user clicks the chip, a call is made from Java to the native
side, passing the stored amount. If the amount is valid, the native
side calls back to Java with the `BnplIssuerContext` to display the
BNPL issuer selection screen. The logic for handling invalid amounts
will be addressed in an upcoming CL.

Video:
http://screencast/cast/NjU0ODk2NzIzMDc5OTg3MnwwZTBkZTUxYi00NQ

Design: go/autofill-clank-bnpl

Bug: b:438784702, 430575808
Change-Id: Idc63d803851fc9892ddc4be35762b04e7c690001
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7014892
Commit-Queue: Viplav Kadam <viplavkadam@google.com>
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Norge Vizcay <vizcay@google.com>
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Cr-Commit-Position: refs/heads/main@{#1531586}