[BNPL][Clank] Handle BNPL issuer suggestion selection
This CL adds the logic to handle a BNPL issuer suggestion selection.
When a user selects a BNPL issuer suggestion on the bottom sheet, we
trigger a JNI call to `TouchToFillDelegateAndroidImpl`, which invokes
the `selected_issuer_callback` with the selected issuer. This callback
triggers `BnplManager::OnIssuerSelected`, which continues the BNPL flow.

Design: go/autofill-clank-bnpl

Low-Coverage-Reason: COVERAGE_UNDERREPORTED
Bug: 430575808, b:438785390
Change-Id: Ied920007f0209e2c64ea229757e1eababe993135
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7018890
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Fernando Ramirez <ferny@google.com>
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Viplav Kadam <viplavkadam@google.com>
Cr-Commit-Position: refs/heads/main@{#1531590}