Clean up IntersectionObserver scroll optimization bookkeeping
LocalFrameView keeps a running total of the magnitude of all scroll
offsets applied to a page, and applies it to all
IntersectionObservations each time the IntersectionObserver algorithm
runs. This happens in two places during a BeginMainFrame: 1)
LFV::ComputePostLayoutIntersections, which runs zero or more times
during the style/layout/prepaint loop; and 2)
LFV::UpdateViewportIntersectionsForSubtree, which runs once during
post-lifecycle steps. Both of those methods are recursive over the frame
tree, but the recursion logic is different between them (e.g. they
handle render throttling and fenced frames differently).

The upshot is that the bookkeeping of accumulated scroll offset can be
unpredictable. This CL makes it so that accumulated scroll offsets are
only applied to IntersectionObservations during the post-lifecycle frame
tree walk (which is always more complete than the post-layout version).

Bug: chromium:40914013

Change-Id: Idd89bf6bf42a19aba9641ed82f1a4e407e8268a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7034645
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531594}