[Bookmarks Bar] Update sizing of elements based on UI review
This CL has non-functional changes to the bookmark bar to update some
of the sizing and spacing based on some ad-hoc in-person testing. We:

- Decrease icon size
- Decrease text size
- Decrease ripple padding
- Decrease ripple radius
- Decrease padding on top of bar
- Increase padding around divider

We update the naming of the default chip ripple to be bookmark bar
specific, since we are deviating from what may be the norm and do not
want others to reuse the resource by default.

Bug: 452714846
Change-Id: Ib555e75d0586b491fc0715c98fa76f682a9b4550
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7050123
Reviewed-by: Brandon Wylie <wylieb@google.com>
Reviewed-by: Eleanor Lee <eleanorlee@google.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1531620}