Modify pref hash browser test to use browser's default search manager
The PrefHashBrowserTestDefaultSearch.SearchProtected creates it's own DefaultSearchManger to test the HMAC protection logic. This is unnecessary as the browser test creates it's own DefaultSearchManger which initializes with the attacked preferences. This CL modifies the test to use the browser's DefaultSearchManger.

Note that to improve the DSE mirror metrics accuracy, we will start clearing the mirrored DSE pref if an HMAC check has cleared the original DSE pref. This detection and clearance happens at initialization of the DefaultSearchManager. This would have caused this browser test (before this modification) to fail as it expects a pref mismatch from both manager initialization. The histogram has been updated to expect one log.

Bug: 450082818
Change-Id: Ieeb62203c8f0a40b06c9e2f8bfbfc157fff015c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7056422
Reviewed-by: Zack Han <zackhan@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Fikre Mengistu <fikremengistu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531648}