[Harden Prefs] Store paths of reset preferences instead of values
This CL re-implements the preference reset tracking feature in a more
stable way. A previous CL introduced a mechanism to store the values of
preferences that were reset by the PrefHashFilter. It introduced a fatal
crash when attempting to call base::Value::Clone() on a preference
object that was in a corrupted state.

The prefs.tracked_preferences_reset preference is changed from a
DictionaryPref to a ListPref. The logic in TrackedAtomicPreference and
TrackedSplitPreference is updated to append the string paths of reset
preferences to this list.

Bug: 448760716

Change-Id: I86d6154d3fa785ee0ce6f51d2929fb819d59418f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7042144
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Zack Han <zackhan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531665}