[ntp][realbox] Fix OGB disappearance in AIM with a background image
This CL fixes the issue of OGB disappearance in AIM with a background
image.

This issue occurs when the following conditions are met:

- A background image is set in chrome
- Realbox Next is enabled
- The scrim is applied in AIM

When these are met in the original code, the z-index of the stacking
context of OGB becomes -1, resulting in the context having a lower
precedence than the background image (which has z-index set to auto).

The change has been checked manually:

Realbox Next Enabled:
https://screencast.googleplex.com/cast/NTgwNjI2MzQ1MDQ2ODM1MnxlMTgwNWI0YS0zMg

Realbox Next Disabled:
https://screencast.googleplex.com/cast/NTU4MzI5NDY1MjQ4MTUzNnw2YTU0N2Y0MC1iMA

Fixed: b:452993633
Change-Id: I3e403b2a015beef4af69ff58d4ac93918045c223
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7056573
Reviewed-by: Riley Tatum <rtatum@google.com>
Commit-Queue: Sho Mizoe <shomizoe@google.com>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531680}