media/gpu/chromeos: fix generic DMABuf mapper and test helper
The number of planes in a layout is usually 2 for NV12. This fixes a
crash where the layout planes were accessed out of bounds.

This avoids the out of bounds access of the layout planes by moving the span creation to loops that are limited by the number of planes.

Bug: 338570700
Change-Id: Ia20d88fa660b195262a2f309d5a8ab3218374622
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7055989
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Auto-Submit: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Eugene Zemtsov <eugene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531684}