Do not synchronously generate IntersectionObserver notifications
Prior to this CL, we would synchronously generate a not-intersecting
notification when a target element was removed from the DOM. That
doesn't match the spec -- there is no affordance for generating
notifications outside of the "update the rendering" steps.

This CL removes all the synchronous forced computation of intersections.
It makes IntersectionObserverController continue tracking disconnected
targets until they have a chance to issue a "not-intersecting"
notification during a rendering update.

Bug: chromium:40914013
Change-Id: I332dc4e95dc9cffbe7d05ec63eb3cd6122c431a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7038511
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531750}