Set Android download params for embedder model
This change configures download scheduling parameters for the passage
embeddings model on Android.

To optimize for mobile conditions, the model downloads are now
scheduled with the following parameters on Android builds:
- Priority: HIGH
- Battery: BATTERY_SENSITIVE
- Network: UNMETERED

This ensures that the model is only downloaded when the device has
sufficient battery and is connected to an unmetered network, minimizing
user impact. No changes were made for other platforms.

Bug: 451969299
Change-Id: I6de2f290c9d8cbe5b5a0be717ca723a8e2e77e57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037117
Reviewed-by: Mike Wittman <wittman@chromium.org>
Reviewed-by: Elias Klim <elklm@chromium.org>
Commit-Queue: Judith Hemp <hempjudith@google.com>
Cr-Commit-Position: refs/heads/main@{#1531809}