Refactor non-trivial callers of chrome::NewTab to specify context
chrome::NewTab() logs to the "Tabs.NewTab" histogram with the value
NewTabTypes::NEW_TAB_COMMAND which is supposed to capture the situation
when the user makes a new tab from the three dot menu, or perhaps some
other menu. But there was a change for the new tab button to also call
chrome::NewTab() on click, which affects this histogram.

This CL fixes this issue by having callers of chrome::NewTab specify the
context of how the new tab is being created. For now, it takes a default
value as there are hundreds of callers of this function, mostly in test
files. But the important callers are changed to specify a value.

Bug: 450329018
Change-Id: I29a9764b8139f68d97f9a3c941af8eafbf004c29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7023721
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Dominic Austria <dominicaustria@google.com>
Reviewed-by: Alison Gale <agale@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532042}