YourSavedInfo: Update Autofill label in 3-dot menu
This commit changes the label in the 3-dot menu (wrench menu) of the
Autofill menu item to display "Your saved info" when
`kYourSavedInfoBrandingInSettings` feature flag is enabled.
Otherwise, it defaults to the original "Passwords and Autofill" label.
It also changes the icon next to it.

Screenshots:
- when feature flag is off: https://screenshot.googleplex.com/3qcfxwnHYc5oDim
- when feature flag `YourSavedInfoBrandingInSettings` is on: https://screenshot.googleplex.com/4ZHT8wX33vwergd

This change is part of the "Your Saved Info" rebranding effort.

Bug: 452603303
Change-Id: I3643001e3f144702e19da8c1437fa532c08c754a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7053331
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Julia Sobiech <jsobiech@google.com>
Commit-Queue: Ireneusz Szulc <ireneuszszulc@google.com>
Cr-Commit-Position: refs/heads/main@{#1532107}