[rustmojo] Establish preliminary structure for the Rust-Mojo interface.
The idea here is:
* the "mojo_c_system_bindings" are the bindings autogenerated by bindgen
* the "mojo_ffi" target is a thin wrapper around those bindings,
  exposing whatever the rust system API will need
* "mojo_rust_system_api" is the system API that will be used by all
  higher layers in the stack

I cribbed from the deleted code for a bit of this, but mostly this CL
exists to get the structure in place (e.g. get a bunch of the BUILD.gn
futzing done upfront); I'll be iterating on top of this with a more
proper example, setting up unit tests, and so on.

BYPASS_RECITATION_REASON=The bulk of this CL is re-adding in code that was previously in the Chromium tree and thus is not problematic to add back in.

Change-Id: Ieaa9f60adc992e597d53ea56c7ca9117ae9053ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7046608
Commit-Queue: Julia Hansbrough <flowerhack@google.com>
Reviewed-by: Devon Loehr <dloehr@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532127}