[PEPC] Implement the accuracymode attribute
The geolocation element supports an `accuracymode` attribute to mirror
the proposed `accuracymode` attribute in the approximate location
proposal:
https://github.com/WICG/PEPC/blob/main/geolocation_explainer.md#attributes-and-events
https://github.com/explainers-by-googlers/approximate-geolocation?tab=readme-ov-file#positionoptions-dictionary

This CL adds this attribute and ensure it updates the text accordingly
(previously handled by the preciselocation attribute). It also ensures
the preciselocation attribute does not have any effect on the
geolocation element.

There are no changes to the API call yet since the `approximatemode`
dictionary option is not yet supported.

Fixed: 448812228,448812824,448811197
Change-Id: Ic6beb77e3423d1732f52fc395b851f4322272d78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7018332
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532156}