[ios] Forward Clipboard reads to DataControlsTabHelper
This patch is part of the effort for implementing Clipboard protections
for Chrome Enterprise on iOS. The protection logic requires detecting
when some content was pasted (or read from the clipboard) into the web
content. The patch forwards the pasted/clipboard read signal to the
Chrome iOS component in charge of Clipboard protections.

Bug: 448120059
Change-Id: I2f0f2f17b30c8ff89afa9774681811da8b341900
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7055425
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Ernesto Izquierdo Clua <eic@google.com>
Cr-Commit-Position: refs/heads/main@{#1532246}