[iOS] Fix DownloadNativeTaskBridge cancel method
DownloadNativeTaskBridge which owns and acts as delegate for a
`WKDownload` contains a `_startDownloadBlock` passed through
`download:decideDestinationUsingResponse:suggestedFilename:completionHandler:`
which must be called before it is deallocated. This block is described
by the documentation as "A closure you invoke with a destination file
URL to begin the download, or nil to cancel the download.". This CL now
passes nil to the block inside `-[DownloadNativeTaskBridge cancel]`
since cancelling is the intent, instead of passing a random file path
and then later removing that path. This makes the code simpler and
solves an issue where a new download task fails when it replaces a task
that was still pending start.

Bug: 452302856
Change-Id: I511f692abdf1a78e3dd16acdae196ad1429e2fc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7054522
Auto-Submit: Quentin Pubert <qpubert@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532286}