[Predictable App Updates] Integrate with generated icon fix infra
The app updating algorithm has a couple of issues because it did not
integrate with the generated icon fix infrastructure, like:
1. The fix to crbug.com/40836693 wouldn't work.
2. The is_generated_icon_ field of the web app would not be in sync
with what the actual icons of the web app are, so there could be
generated icons with the field set to false, leading to data being
out of sync for those who are reading that field.

This CL fixes all of that and verifies that it all works with the
existing sync tests that install web apps from sync and verifies the
generated icon fix behavior works as expected.

1 is fixed by ensuring that certain apps can update their identity
within a particular number of days if they have broken icons installed
via sync.

2 is fixed by ensuring that we ONLY set the generated icons field in
the web app when we actually write the new icons to the web app. In
case we write the old/existing ones, the app's generated icon state
is copied over to the install info pre update.

Fixed: 445180355
Change-Id: I7f5ee468c37896b304fab7bde764130e311cc9fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7050060
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532289}