[iOS] Prevent crash in HomeCustomizationBackgroundCell caused by a nil logo mediator
A crash occurred in HomeCustomizationBackgroundCell when setting up
layout constraints using the logo mediator. In some cases,
searchEngineLogoMediator.view returned nil.

When a nil view is used to create a constraint (for example,
logoView.widthAnchor...), the resulting NSLayoutConstraint is also nil.
Adding that nil constraint to the constraints array triggers an
NSInvalidArgumentException and crashes the app.

To fix this, an early return was added to verify that both
searchEngineLogoMediator and its view property are non-nil before
creating constraints or accessing the view.

Fixed: 452898329

Change-Id: I5d591e5404e85fbeecdc8fff3d6ddf3b28fc3a1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7059350
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Commit-Queue: Pier-Alexandre Bouchard <pabouchard@google.com>
Cr-Commit-Position: refs/heads/main@{#1532305}