[ios] Re-enable test testOpenSearchWidgetWithProfileSwitch
This change makes sure that test testOpenSearchWidgetWithProfileSwitch
is run only when feature kSeparateProfilesForManagedAccounts is enabled.

It also makes sure the omnibox is unfocused before switching profile
(there is a dangling raw_ptr when switching profile with the omnibox
focused), this will be fixed separately since the purpose of this test
is not to test that.

Fixed: 450574249
Change-Id: I3246e9d978244e3f96c7971aa5614266d0d56295
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7036579
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Federica Germinario <fedegermi@google.com>
Commit-Queue: Federica Germinario <fedegermi@google.com>
Cr-Commit-Position: refs/heads/main@{#1532322}