[iOS] Remove reference to closed bug 40273505
This bug has been closed because the promos manager now does correctly
track whether there is a displayed promo. Thus, we no longer need to
differentiate the first promo shown from the rest.

Before, if the app restarted while a promo was already displayed and
there was a promo forced in experimental settings, we'd try to display
the forced promo, which caused issues. Now, the code doesn't even reach
there because the manager knows a promo is already being displayed.

Bug: b:40273505
Change-Id: Ie91b2740b163b660606005f7983e3e8ed451b61d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7055743
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Gayane Petrosyan <gayane@google.com>
Cr-Commit-Position: refs/heads/main@{#1532341}