Correctly track swap chain size on reallocation in SwapChainPresenter
In the failing auto HDR fallback case, we reallocate out swap chain with
RevertSwapChainToSDR, which leaves the content_size_ and
swap_chain_size_ fields in an invalid empty state. This causes issues
with the EarlyFullScreenOptimization feature, which depends on knowing
the swap chain size.

Bug: 448970702
Bug: 449165645
Change-Id: I312f2e5cc351a89fee41f14d9ad22ae9d68dc39e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7054748
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Commit-Queue: Michael Tang <tangm@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1532360}