[iOS]Uses GaiaId type when possible in ['consistency promo']
GaiaId in iOS is currently often encoded as a NSString or more rarely
a std::string. It should ideally be replaced by GaiaId, a c++ class
that contains simply the gaia id (or an empty string to represents the
id missing).

In order for the migration not to break ios_internal, some virtual
method has to be temporarily duplicated, taking both kind of
arguments. In this case, both functions call the other function, and
the classes inheriting this abstract class must implement one of the
two functions.

This is part of the chain
+ crrev.com/c/7003548 (that is splitted)
- crrev.com/i/8634707
- crrev.com/c/7021136
- crrev.com/i/8650938
- crrev.com/c/7021672

This CL was uploaded by git cl split.

Bug: 449195064
Change-Id: I06956f2ad2435cbec322d430f250c985640790c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7025631
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: Menghan Yang <myuu@google.com>
Commit-Queue: Menghan Yang <myuu@google.com>
Cr-Commit-Position: refs/heads/main@{#1532371}