[Masonry] Move grid intrinsic minimum logic to helpers
The grid implementation has special logic for intrinsic minimums that
also should apply to masonry [1].

Move this logic to two new helpers:
- CalculateIntrinsicMinimumContribution() that returns the contribution
size to use for a given item.
- ClampIntrinsicMinSize() which clamps the contribution size, if it
was the min content contribution, by the total track size that the
item spans.

This is in preparation for reuse in masonry. This CL should produce
no functional changes.

[1] https://drafts.csswg.org/css-grid/#min-size-auto

Bug: 343257585

Change-Id: I6b13bf9bafb59aa7b7006486aa46fd346304afd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7013328
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1532392}