[content] add option to ignore a11y input events
WebContents::IgnoreInputEvents suppresses user input to the WebContents.
However this doesn't automatically block accessibility actions that can
interact with the WebContents.

This change parameterizes IgnoreInputEvents with a new `bool
should_ignore_a11y_input` that defaults to false. Callers can now pass
`true` to block all accessibility actions other than the hit test.

Motivation for this CL: The ActorOverlayWebView uses
ScopedIgnoreInputEvents to block input events when the webview is
visible. However the existing implementation of IgnoreInputEvents did
not account for assistive technologies interacting with the WebContents.
This is necessary because input events to the underlying webcontents of
an actuating tab can disrupt the actor task.

This CL adds the first use case of should_ignore_a11y_input=true in the
ActorOverlayWebView.

Screencast:
https://screencast.googleplex.com/cast/NTgwMzIyMTIwNDMzNjY0MHwwYjViODYxOC05Nw

Bug: b:441073485
Change-Id: Ice96ab40f1cf507d97e713833e4846b8e9afb13d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7050137
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>
Commit-Queue: Jacob Stanley <jacobstanley@google.com>
Cr-Commit-Position: refs/heads/main@{#1532457}