Reland "Avoid redundant call to NotifyNavigationStateChanged."
This is a reland of commit 7d1910ab3a702eb2b7a8ac8eb69e13da59a0a5b0

The original change was backed out due to a flaking
`NewTabPageURLOverride` test that was not seeing the location bar text
being cleared. This happens because, when the browser test suite is run
with parallel test jobs, there are sometimes extra focus calls being
made on the omnibox, which calls `OmniboxViewViews::OnFocus` and makes
an extra call to `ResetDisplayTexts()` outside of
`OmniboxViewViews::Update`. That call does not invoke `RevertAll()`, so
the actual text the test checks isn't properly cleared, but by the time
the other `ResetDisplayTexts()` call happens, it will think the text was
cleared out. This results in the test waiting for the omnibox to change
from "about:blank" to "", which never happens. This was not an issue
before this change because the redundant call to
NotifyNavigationStateChanged happened before the focus call, ensuring
that the display text was properly cleared out in both places.

The extra call was added in response to this DCHECK bug:
crbug.com/40551417. However, that DCHECK was made less strict in
crrev.com/c/3888302, and the extra call has a comment questioning
whether it's needed. This CL fixes the test flake by removing that extra
call altogether. This is somewhat risky since there is no test in place
for the original DCHECK issue, but I was not able to reproduce the crash
with the steps listed in the bug.

Original change's description:
> Avoid redundant call to NotifyNavigationStateChanged.
>
> During RendererDidNavigate, there are often two identical calls to
> NotifyNavigationStateChanged, both of which are time consuming. This CL
> ensures that exactly 1 call will be made when the function returns. It
> accomplishes this via a class scoped to RendererDidNavigate that invokes
> the call to NotifyNavigationStateChanged in its destructor. Where the
> function used to be called unconditionally, it now "arms" the new
> ScopedDeferNotifyNavigationStateChanged class (if the class was passed
> in as a parameter to the function). The class is only passed in to
> helper functions from RendererDidNavigate; all other code paths will not
> have this passed in and will continue to unconditionally send the
> notification.
>
> This change is guarded behind the following killswitch:
> SkipRedundantNavigationStateNotification
>
> This CL also fixes an issue on Android where the tab URL tracked by the
> LocationBarModel and the CustomTabToolbar can get out of sync for
> trusted CDN pages. This is a bug that was exposed when the extra
> NotifyNavigationStateChanged call was removed.
>
> Change-Id: I7f17af7dacd87e3fb42fb13a97e949479d2e803f
> Enabled-by-default-reason: killswitch
> Bug: 441518347
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6538910
> Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
> Reviewed-by: Matthew Jones <mdjones@chromium.org>
> Commit-Queue: Liam Brady <lbrady@google.com>
> Cr-Commit-Position: refs/heads/main@{#1511173}

Bug: 441518347
Change-Id: If0ae2990dfd4aece0fefea1c44530b0be46b6acc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6944687
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532460}