[IOS] Fix Filled card suggestion metadata wrongly logged
When the CVC authentication UI is presented, it was triggering a new
Autofill suggestion query. This new query would overwrite the logging
context from the original card selection, leading to the loss of
metadata.

In this CL, the CardUnmaskAuthenticationCoordinator now calls
DidCoverWebContent when the CVC prompt is shown and DidRevealWebContent
when it is dismissed. This ensures the webState->IsVisible() property is
correctly updated, allowing FormInputAccessoryMediator to reliably check
this state and suppress the unwanted suggestion query.

Bug: 442876403

Change-Id: If30b2f896e0ed691833b8b444b4fc978bb1f72f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7030233
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Siyu An <siyua@chromium.org>
Commit-Queue: Yiwen Qian <yiwenqian@google.com>
Cr-Commit-Position: refs/heads/main@{#1532486}