Webium: Fix crash in Document Picture-in-Picture
Currently in Webium, Document PiP will pass the checks in the
renderer process to create the new window, but fail in the browser
process during RenderFrameHostImpl::CreateNewWindow. This causes the
renderer process to crash.

The cause is due to using GetParentOrOuterDocumentOrEmbedder, which
"escapes" the GuestView/GuestContents, and returns "false" for our
top-level check. Instead, we should use IsOutermostMainFrame, which
most closely matches the desired effect of the check, and aligns
with the renderer process check. Notably in Webium, it will return
"true" for the top-level check within the tab's GuestContents.

Fixed: 452079354
Change-Id: I410d2fcff978376226f120e9641eb8535c80fb36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7047382
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Keren Zhu <kerenzhu@chromium.org>
Commit-Queue: Nic Williamson <champnic@microsoft.com>
Reviewed-by: Tom Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532488}