[Masonry] Implement intrinsic minimum logic
The grid implementation has special logic for intrinsic minimums that
also should apply to masonry [1]. This CL implements this logic for
masonry.

In masonry, we need to store the intrinsic min contribution on the
virtual item since we want the max between all items it represents. The
logic also depends on the tracks an item spans, which we don't know when
creating the virtual items. This required storing 4 new vars on the
virtual item, listed below. Since we now have 5 vars related to virtual
item contributions in GridItem, create a new struct to hold these called
VirtualItemContributions.

1. `intrinsic_min_assuming_track_placement` - which holds the max
intrinsic min contribution between items assuming they all span tracks
that force the contribution to the automatic min.

2. `intrinsic_min_ignoring_track_placement_unclamped` - holds the max
intrinsic min contribution where we assume the span does not force the
contribution to the automatic min. In this case, some contributions will
need to be clamped. This holds the max contribution for items whose
contribution doesn't need to be clamped.

3. `intrinsic_min_ignoring_track_placement`, same as #2, but for items
whose contribution may need to be clamped.

4. `min_clamp_size` - the max sum of margin/border/padding/baseline shim
for all items that contribute to #3. The contribution should not be
clamped past this point. Note that this won't be 100% equivalent to grid
clamping, as grid clamping will apply this min clamp size to the item it
applies to, whereas we are storing the largest such sum. This should
produce close enough results to grid without overflow.

These pieces of information give us enough to produce similar results to
grid's implementation of intrinsic minimums.

We also now call CacheInitializedSetsGeometry() before getting the
virtual item contributions because we need this to be able to tell if
any of the tracks a virtual item spans are definite.

ContributionSizeForVirtualItem() was also moved to be part of
MasonryLayoutAlgorithm so that it can have different inputs than
ContributionSizeFunctionRef.

Note that there are still a chunk of masonry test failures from CL:7014447.
A small subset are because we don't incorporate baseline shims yet. One
had the correct sizing, but positioned the abspos elements at the wrong
start offset. The other subset need to be investigated. Some may be actual
failures, and some may just be expected diffs between grid and masonry -
I'll spend time going through those in follow up change(s) in more detail.

[1] https://drafts.csswg.org/css-grid/#min-size-auto

Bug: 343257585
Change-Id: I113905e4e712ea71daf8bfb5bc679bd4ac49ca12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7033351
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1532600}