Optimize map lookups in `WebContentsImpl`.
This change refactors map operations to reduce redundant lookups. It
replaces separate `contains()` and map access (`[]` or `erase`) with a
single `find()` call and iterator usage. Additionally, it uses the
return value of `insert()` to check if an element was added to
`fullscreen_frames_`, avoiding a redundant `Contains()` check.

Bug: 452950815

Change-Id: Ia1fa079d6b3e76ec1c58b9a873a6873d7944f803
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7047392
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Jeffrey Yu <yuje@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532618}