[BrowserControls] Integrate SHOWING / HIDING visibility types
TopControlsStacker is capable to drive height update animations. This
requires the layer to coordinate accordingly, setting their visibility
types when they are requesting height changes and they are expecting an
animation to run.

Note: This current patch do not work when browser controls's minHeight
is reducing while other control layers are scrolled off. It will be
fixed in follow up work.

In order for the stacker to communicate back to the layers when the
animation finishes, the TopControlLayer added a |reachRestingPosition|
boolean. The layer can adjust its visibility once they are notified they
have reached the resting position.

Bug: 417238089
Change-Id: Id24c3a1c5e20ee47f763fdffcb9f0a8083369374
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7050767
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532650}