[Bedrock] Remove Views access from Browser[Window[Interface]]
This removes accessors for LensOverlayView and TopContainer from
Browser, BrowserWindow, and BrowserWindowInterface. This is part of an
effort to remove the dependence of all of those classes on Views.

These can already be retrieved via BrowserElementsViews with the correct
ElementIdentifier.

In addition, in several places in the code,
BrowserWindowInterface::TopContainer() was being used as a proxy to
retrieve the browser widget. The only problem is that - especially in
Chrome Next due to some reshuffling of views - the top container might
be on another widget in Immersive Mode on ChromeOS or Mac. It's not
clear if these discrepancies were in error or intended.

TODO: before submitting, determine if these retrievals are in error, and
if so, simplify the code to only use the actual browser widget.

Bug: 453638114
Change-Id: I6fae211005b733c113f5e33d78ce532200ca3502
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7059373
Reviewed-by: Dan Harrington <harringtond@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532661}