[owners] Remove asully@chromium.org from chrome/browser/sync_file_system/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/asully@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I509cb368de4dbbdabe333de75282499909f7731c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7059782
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1532709}