[//chrome] Transition VideoStreamView to call PCVR::PaintOOPR()
We are incrementally hardening PCVR::Paint() to be able to verify that
its passed-in raster context provider is either (a) null or (b) supports
`gpu_rasterization`. This is not actually the case at this callsite, as
it uses the context provider constructed here [1]. PCVR::Paint() is a
no-op if passed a VideoFrame with a SharedImage and *not* passed a
context provider that supports GPU rasterization [2]. Hence, we hoist
that check to the callsite here to be able to port the call here to be
on PCVR::PaintOOPR().

Note: It's not clear whether the client-side behavior here is expected
or not, but in any case, this CL doesn't introduce any functionality
change in that flow.

[1] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/compositor/viz_process_transport_factory.cc;l=514-556;drc=33aa41994c585274ba016ef43a3b8baa3c6334a0
[2] https://source.chromium.org/chromium/chromium/src/+/main:media/renderers/paint_canvas_video_renderer.cc;l=899-909?q=PaintCanvasVideoRenderer::Paint&ss=chromium

Bug: 391648152
Change-Id: Idf8b55acf04afc2cf582b566a040a9f2f724046e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7054544
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532765}