[ozone] Update RegisterSysmemBufferCollection to use SIFormat
Update RegisterSysmemBufferCollection in VulkanImplementation and its
overriden method in ozone/flatland to use SharedImageFormat instead
of BufferFormat. Also plumb it from its callsite in SIFactory all
the way to where it is used in FlatlandSysmemBufferCollection, and
update its members to talk in terms of SharedImageFormat.

Bug: 450326410
Change-Id: I1dd10af6fced4d403b9d66bc9d360bbdb28192f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7063431
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532909}