Store all seed info in proto format in the Seed Files.
Instead of only storing the seed data in the Seed File and the rest of
the seed info in Local State, store all the related seed-info along with
the seed in a proto-serialized and gzip compressed in the Seed Files.

Also added new tests to cover this new flow. Note that now some of the
checks will be done when reading the seed from file on startup, instead
of when calling LoadSeed().

Note: we also plan to change the compression format, but since the
change was already quite large, it will be done in the next CL.

Bug: 417138763
Change-Id: I5ed58ca090e1ba93ff738b9e8b0f9a2f4b532c33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7010852
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Ramon Cano Aparicio <rcanoaparicio@google.com>
Cr-Commit-Position: refs/heads/main@{#1532913}