Disable cardholder name in save card bubble when loading
The cardholder name fix flow on desktop has a textbox to allow the user
to change the name that will be used as their Google Payments account's
legal name. In the past, accepting the dialog made the entire bubble go
away, but now the Loading and Confirmation Project allows it to stick
around even when the save is currently happening. During this time, the
textbox is editable, even though the save is already underway!

If the user edited the textbox, it would have no effect on the save but
could be confusing to the user. To fix that, this CL just disables the
textbox when the loading state is entered.

Demo: https://screencast.googleplex.com/cast/NTA5OTQ0NzQ0Mzc4MzY4MHw4YWJjODgxMC0yZA (starts at 0:20)

Bug: 447170347
Change-Id: I30603c2d3f62eb7165aa0b18b892aacb7857367a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6989429
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Olivia Saul <osaul@google.com>
Cr-Commit-Position: refs/heads/main@{#1532935}