Add DeveloperToolsPolicyChecker
This CL introduces `DeveloperToolsPolicyChecker`, a new KeyedService
responsible for checking whether DevTools is allowed for a given URL
based on enterprise policies.

The checker uses `URLBlocklistManager` to evaluate the URL against the
`DeveloperToolsAvailabilityBlocklist` and
`DeveloperToolsAvailabilityAllowlist` policies.

This will be used in a follow-up CL to restrict DevTools access based on
the URL of the inspected page.

Unit tests for the checker and its factory are also included.

Bug: 442892562

Change-Id: Ic28ff3173dbd54f7c597a8645523acfa5c11658a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037041
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Michał Kaczmarczyk <mickaczmarczyk@google.com>
Cr-Commit-Position: refs/heads/main@{#1532946}