Remove current tab synchronously
This CL allows tab to be closed without delay if this tab is closed
during minimization.

This prevents a race condition where the delayed runnable could execute
after the tab was destroyed, causing it to reappear on the next startup.
Removing the delay ensures the tab is closed correctly.

Activity.isFinishing is not reliable yet which always returns false at
that stage and so is unable to tell if the app will closed.

HARD_TO_TEST: needs the app to be closed

Bug: 450560278
Change-Id: I0617bc006491ee7b9c018300f5bf3a4cd9dd8023
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7056579
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Lijin Shen <lazzzis@google.com>
Cr-Commit-Position: refs/heads/main@{#1532992}