[ios, tabgrid] Polish drop transition from dropped on cell to new group
Upon a drag and drop to create a new group, the dropped on cell is
highlighted. To perform a seamless transition, upon configuring the
new group cell, highlight it so the group cell has the same shrunken
transform as the dropped on cell upon replacement. Then animate the
reset. Some properties are added to handle the situation where the
final index path of the group cell is different from the dropped on
cell. This happens when the dragged cell's index path is smaller than
the dropped on cell.

Video: https://screencast.googleplex.com/cast/NDc4NTQzNjg5ODYyMzQ4OHxhMzM5NjAwMC05Mg

Bug: 450613202
Change-Id: Id2ac0091def3bb405da08dcd987d1c17725cd498
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7046949
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532994}