Fixing a crash in TabModal positioning in MacOS Immersive Full Screen
The toolbar view isn't parented to the browser view in MacOS immersive full screen mode. As a result we fail with a CHECK when converting to browser view coordinates. To fix this we will instead convert it to coordinates of the widget. This addresses the crash issue seen.

Fixed: 453930466
Change-Id: Ie5bc76b23edc5289ee6aa47d5a82c7c396d335db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7066276
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533015}