Keep the most promising groups of UnimportantFramePolicy in testing config.
With this CL, we only keep the IsolatesPriority and All groups of the
UnimportantFramePolicy study in the testing config. So far, field
experiments reveal that IsolatesPriority is the most impactful
standalone group, and there are signs that All provides a little bit of
additional impact. We'll run a field experiment with only All and
IsolatesPriority (25% each) to validate the hypothesis that All provides
a little bit of additional compared to IsolatesPriority.

Bug: 40942028
Change-Id: I3a2974b41d5e9d5db621193bfd70dd06b3961f1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7063662
Commit-Queue: Olivier Li <olivierli@chromium.org>
Reviewed-by: Olivier Li <olivierli@chromium.org>
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533204}