[ReduceScreenSize] Drop this experiment.
As per https://privacysandbox.google.com/overview/status, we're not
going to ship this. Let's drop it from the codebase.

As this is the final intervention which was gated on the never-shipped
`kIncognitoFingerprintingInterventions` flag, we'll remove that flag
here as well.

This CL contains reverts for the patches landed in the
`reducescreensize` hashtag (see [1]).

[1]:
https://chromium-review.googlesource.com/q/hashtag:%22reducescreensize%22

Bug: 408932088
Change-Id: I02b7ff78c988b68a24743befe93a638fd34bbb43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7062630
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Thomas Nguyen <tungnh@chromium.org>
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533416}