SmartCard: Fix crash on opening Chrome settings in guest mode
This remediates the crash by just enabling the
SmartCardPermissionContext service in guest profiles as well. Note that
this is probably necessary, as most likely Web Smart Card API will not
ever be used in guest mode. That being said, another way would be to
disable also all the UI elements related to it in guest profiles, which
is a much more complex endeavour (although can be considered as a future
improvement: crbug.com/454152416).

Bug: 40257657, 454152416
Change-Id: I45ca7fd5d0aadb76a0beae45d07e7cd7717bf62d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7072034
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Commit-Queue: Zgroza (Luke) Klimek <zgroza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533530}