[sync] Update URL for Sync Dashboard (behind disabled flag)
...which is currently referred to as the account data dashboard.

The new URL is considered more modern but otherwise serves the same
content as the old one.

The change is guarded behind a newly-introduced feature flag, which is
disabled by default for now. Follow-up patches plan to enable the
feature by default, and use it as a kill switch.

Bug: 447134784
Change-Id: I6caae8c67aeebc780df4ae810e36ffc119296c93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7052385
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Tomek Jurkiewicz <tju@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Duong Dac <ddac@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533541}