[TreesInViz] Improve surface damage tracking consistency
Before this change, we have a path in surface damage tracking that, in
the case where a SurfacePropertyChanged() but not
AncestorPropertyChanged(), we set the damage to the render surface's
content_rect.

If AncestorPropertyChanged then we were using
damage_from_leftover_rects, but not content_rect.

This seems inconsistent, and it adds a strong dependency on ancestor
EffectNode::effect_changed to be one way or the other, that are
difficult to reason about and caused a tab hang bug in TreesInViz.

This change merges all of the surface damage sources, which seems safer
and easier to reason about, and fixes the TreesInViz issue.

CHECKS are added to TreesInViz code to validate damage consistency.

Bug: 450022697, 449909527
Change-Id: I4847eae6e7479d272665d1d55331a0ab6a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7068008
Reviewed-by: vikas soni <vikassoni@chromium.org>
Commit-Queue: Victor Miura <vmiura@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533603}