[WebAuthn] Prevent null icon field in password responses on Android
There is currently a crash with immediate mediation when a password
credential is selected on a device using CredMan. The `icon` field
cannot be null in the mojo struct. This adds an empty but non-null
URL for it.

Bug: 433543129
Change-Id: Ie636902930ee063d01d9d2b47a3dcafdce95cbd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7068373
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Auto-Submit: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533609}