[Browser Frame] Handle caption buttons on leading edge in Linux
This handles the case where buttons have been moved to the leading
edge via GTK settings.

BrowserFrameViewLinux never had an implementation of
CaptionButtonsOnLeadingEdge(); this adds one. It also handles leading
buttons when adding the exclusion for the app icon in WebApps (though in
general this is already not placed well on Windows or Linux).

Bug: 453919397
Change-Id: Ied37ea9a1b79c96fda19995487551d91d06e0097
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7073391
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533650}