Bypass service worker for subresources with synthetic responses
When a page is loaded using a synthetic response from a service worker,
subresource fetches for that page were still being intercepted by the
service worker. This interception is unnecessary and adds overhead.

This change introduces a feature param to optimize this behavior. When
the feature flag is enabled, the browser process instructs the renderer
to bypass the service worker for subresource loading if the main
resource was served with a synthetic response and its dry mode.

This change is a temporary one to run the ablation study of subresource
interceptions by synthetic response. Otherwise the synthetic response
feature will be mutually exclusive with subreosurce fetch handlers or
other static routing rules in the future.

Change-Id: Idafa53b33631e629184f2d27786bf40e5a3d307f
Bug: 447039330
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7065556
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533680}