Replace DCHECK with DumpWithoutCrashing for invalid histogram arguments.
When `Histogram::FactoryGet` or `LinearHistogram::FactoryGet` are called with invalid arguments, trigger a `DumpWithoutCrashing` with a crash key instead of just a DCHECK. This allows monitoring of these events in production builds, but silences the errors from fuzzers.

When such an error happens, it could be due to memory corruption or a compromised renderer, or it could be a coding bug where there's two call sites calling a histogram with different args. The latter is useful to detect and route bugs to the histogram code author.

Bug: 451448204, 451649552
Change-Id: I8770113689f77e2593f658da877432ddc283749d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7073831
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Roger McFarlane <rogerm@chromium.org>
Auto-Submit: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Roger McFarlane <rogerm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533692}