mojo: Rename MojoLinuxChannelSharedMem feature to MojoUseEventFd
This is more of a rebranding than anything else. Originally
"MojoLinuxChannelSharedMem" reflected how it enables communicating over
shared memory on the mojo::core::Channel layer. At the time this was the
only way for mojo to use shared memory for communication.

These days, however, ipcz uses shared memory out of the box, so
emphasizing 'SharedMem' in the feature name became confusing.

The 'Linux' part of the name refers to the fact that this feature uses a
feature (eventfd) specific to the linux kernel, so the feature makes
sense only on systems running with it: Android, Linux, ChromeOS.

The new incarnation of the experiment is going to run on Android-only,
so this reference to "Linux" may also be confusing.

Bug: 388531132
Change-Id: Ic4aa27dadec25f1d5c065325e15aa941adf913b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7047037
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533763}