Support lookup of tests by name
Adds --name to search for and runs a test by name.
Also changes the behavior of positional parameters to autotest.py.
If a parameter looks like a file or directory, it is treated like
a test file/directory. Otherwise, it is treated like a --name
argument.

Change-Id: I7075a30a0c6879ffadeb3ef71401d0d36a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7064792
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Dan Harrington <harringtond@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533821}