[bookmarks] Fix bookmarks widget empty folder message load delay
Move the up navigation outside of the bookmarks list in order to enable
using setEmptyView to instantaneously load the empty folder message in
the widget. This circumvents a bug that was happening where the empty
folder message would lag behind. Eg. if a user navigated into a folder
and immediately navigated out of the folder, the empty folder message
could potentially render on top of the bookmarks of another folder and
accept click actions instead.

In addition to moving the up navigation item (folder + back arrow) from
the bookmarks list items is that it is now pinned to the top of the
bookmarks widget. This is an improvement to the previous implementation
where the up navigation item would get scrolled away if user had many
bookmarks and had to scroll down to find the one they were looking for.

Before: http://screencast/cast/NTU2NDMzNzM4ODI1NzI4MHw2MzgxM2ZkYy1kZg
After: http://screencast/cast/NTQwNzU1MTIxNzcyOTUzNnw2YjQxOTJkMS0zMg

Bug: 441959734
Change-Id: I8e5a628a751c9c5450f9fdd2f798c3cefff73e95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7041542
Reviewed-by: Brandon Wylie <wylieb@google.com>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Evan Luo <evanluo@google.com>
Cr-Commit-Position: refs/heads/main@{#1533842}