[iOS] maintain the height of the effect view from the keyboard accessory
Before the fix:
https://screenshot.googleplex.com/BK6ZUaRZqirGPH5
https://screenshot.googleplex.com/8wEEhaBoPMgLTLR

After the fix:
https://screenshot.googleplex.com/A2DXkLN9y4wt7qD
https://screenshot.googleplex.com/AefFE8pzRWHkb63

Note:

The main problem of 442607855 is the keyboard accessory view covers the content beneath it.

This CL is part of the fix. It prevents the effect view to grow by allowing the bottom margin to be flexible, which in turn fixes the vertical alignment of the buttons.

Bug: 442607855
Change-Id: I4d4375d8137731377e7a223502e60361a0f668e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7075414
Reviewed-by: Alexis Hétu <sugoi@chromium.org>
Reviewed-by: Noémie St-Onge <noemies@google.com>
Commit-Queue: Leo Zhao <leozhao@google.com>
Cr-Commit-Position: refs/heads/main@{#1533869}