NTP Promos: Fix "shown" metric reporting when promos are disabled
This refers to the enumerated histogram counting whether promos are
shown or not due to policy-type reasons.

Logging on NTP startup wasn't covering the code path in which the NTP
is able to show promos, but it doesn't because there's no promo
controller. This could be due to either the feature being disabled,
incognito or policy.

Experiment analysis isn't affected by this issue, but it creates
confusion because (1) the control arm logs very few "shown" metrics
compared to experiment arms, despite that it they should match and (2)
experiment arm metric counts are ~10% lower than NTP load metric
counts, when they should match (could be due to promos being disabled
due to incognito or other reasons).

Also, fix the gap in test coverage that allowed an NTP refactor/
cleanup CL to miss this case.

Change-Id: I35926a54fc9f02ae173c343b2e8b5bebb3ec555c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7075533
Reviewed-by: Riley Tatum <rtatum@google.com>
Commit-Queue: Christopher Grant <cjgrant@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533894}