[WebAuthn] Enable timeouts for Immediate CredMan requests
The Immediate Get request timer, which by default aborts the request
if credential enumeration takes more than 500ms, does not currently
work for requests that are handled entirely by CredMan. This is
because the browser cannot know if UI has been shown or not when a
timeout occurs.

This change adds a path through `prepareGetCredential` for immediate
requests. It will now enumerate the credentials in advance of sending
the `getCredential` request which will show UI, allowing an abort
in cases where enumeration is slow.

This CL also does a minor cleanup in `CredManHelper` by removing
redundant cleanup calls during cancellation.

Bug: 433543129
Change-Id: If27bcf04d90df55d877f2fb536d925805a25b0aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7076033
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/main@{#1533897}