omnibox: Speculative fix for AddRowWordsToIndex crash.
Something appears to be corrupting the `word_map_` map in
URLIndexPrivateData, leading to a crash when the tree re-balances itself. It appears this only happens when an empty term is inserted.
Try a speculative fix by skipping empty terms, and add some crash
keys to confirm if this happens for non-empty terms.

Bug: 348617573
Change-Id: I5d98bd48410e0149f89d5ea04260480c60264cfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7067972
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533898}