Disable flaky GeolocationServiceUnitTest.UrlWithApiKey
Even with one fix in place, this test case continues to be flaky.
Disable the test to make the tree green, and give the developer working
on the fix less time pressure to flush out all the issues.

Fix a nit with #include ordering along the way.

Bug: 453817118
Change-Id: I010f7405fe120441b7b18a41d0229c54c0a1a7f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7075641
Reviewed-by: Alvin Ji <alvinji@chromium.org>
Reviewed-by: Zonghan Xu <xzonghan@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533902}