Propagate OverrideState to WebViewCachedFlags
Problem: We currently use synthetic field trials to record metrics for
WebViewCachedFlags. However these metrics are imbalanced because
WebViewCachedFlags only tells us if a feature is enabled or disabled and
does not differentiate between if a feature is disabled as part of the
control group or just disabled because the client is not part of the
study.

Solution: This cl only stores the cached flags if it is overridden ie
the client is actually part of the study. This change should allow for a
more accurate accounting of a client's experiment groups.

Also this cl registers a `WebViewStartupTasksMetrics2` to use this new
mechanism.

Bug: 446603794
Change-Id: I9bcfb76d75c74a3afd092e700782c96c1e34f03e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6943133
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Commit-Queue: Chidera Olibie <colibie@google.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Alex Mitra <alexmitra@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1533925}