[composebox] Update file thumbnail styling in Composebox.
This change updates the background color, padding, height, and spacing
of the file chip. It also adjusts the thumbnail size, adds a background
and border-radius to the thumbnail container, and updates the font size,
weight, and line-height of the title. The favicon within the tab
thumbnail is also resized and given a circular border-radius.
Screenshot: https://screenshot.googleplex.com/8AZ8BDNVdc6Ebng
https://screenshot.googleplex.com/5z9kubATRBuBvoB

Bug: b:452481330
Change-Id: I470733b7ea6172638da863b60b1a3b9b9fd7d852
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7068602
Reviewed-by: Riley Tatum <rtatum@google.com>
Commit-Queue: Dhruv Kathpalia <dhruvkathpalia@google.com>
Cr-Commit-Position: refs/heads/main@{#1533974}