PaintTiming: Use PaintTimingRecord state in LCP Calculator
Currently, the NotifyMetricsIfLargest[Image|Text]PaintChanged] methods
have a bunch of parameters whose values come from the corresponding
PaintTimingRecord, which is also a parameter. This CL uses the state
directly where it can.

This is mostly straightforward conversion, except for the timestamp used
for images for metrics (UKM). Hard LCP uses the first animated frame
time for images and videos, and paint time (which is actually
presentation time...) otherwise. But soft navs always uses the latter.
We should be able to remove this discrepancy once we ship
ReportFirstFrameTimeAsRenderTime (crbug.com/449779010).

Bug: 454082771
Change-Id: I5ff6ce04af87b81c8646d35b74b04f95f6c1b5c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6955664
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534013}