arc: Skip blocking salt read on UI thread
This change prevents a browser crash in developer builds on Reven
that occurs when a user logs in before the ARCVM DLC installation is
complete.

The crash is caused by a blocking file system check (base::PathExists)
on the main UI thread. This happens when OnPrimaryUserProfilePrepared()
calls GetSerialNumberForKeyMint(). If the ARCVM image has not yet been
installed from DLC, this function attempts to read the ARC salt file
(/var/lib/misc/arc_salt) synchronously, triggering a DCHECK
(!tls_blocking_disallowed) and crashing the browser.

This change modifies ReadSaltOnDisk() directly. It now includes a check
to detect if it is being called on the main UI thread. If so, it logs
an error and, in DCHECK-enabled builds, immediately returns an empty
string to bypass the blocking file system calls. This serves as a
temporary fix to prevent the development-time crash.

Returning an empty string is safe because the investigation showed the
salt file is not created or used on Reven devices.

This change fixes a latent bug that would be exposed by
crrev.com/c/6913851

BUG=b:450307883
TEST=unit_tests --gtest_filter=SerialNumberUtilTest.*
TEST=deploy Chrome and verify arcvm availability on Reven
TEST=confirm ro.boot.serialno is available in ARCVM on a Nissa device

Change-Id: I3599649275d97ef295f3b7d20c581b7e8e0b8092
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7043063
Commit-Queue: TingHao Lin <tinghaolin@google.com>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Reviewed-by: Vaibhav Raheja <vraheja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534016}