Add reload-page-with-log.js webtest
reload-with-script.js is currently flaky and is pointing towards a bug
that we have (script may not run at the right time, or potentially it is
a console logs related issue). This adds a new test which checks if we
in general retain logs on reloads.

Bug: 443229243
Change-Id: I741d3813f7f4384e4b0bbc0b84075b6e2a3d4185
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7073711
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534223}