Introduce NavigatorLocks mixin for locks
This patch migrates partial interfaces to interface mixin to align with
the specification.

Originally, this partial interface was introduced as a workaround for a
layering violation between core/ and modules/, since Blink could not
include mixins across those layers.

Now that the IDL compiler supports cross-component mixins[1], the
workaround is no longer needed, and this patch replaces the partial
interfaces with interface mixin to follow the specs[2].

This CL has no behavior changes.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/6968299
[2] https://w3c.github.io/web-locks/#navigatorlocks

Bug: 453570181
Change-Id: I791672adae7939f3384e104bc06119addd729d6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7074371
Reviewed-by: Fergal Daly <fergal@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Suyeon Ji <zeesuyeon@gmail.com>
Cr-Commit-Position: refs/heads/main@{#1534244}