Introduce WindowLocalStorage/WindowSessionStorage mixins for storage
This patch migrates partial interface to interface mixins to align with
the specification.

Originally, this partial interface was introduced as a workaround for a
layering violation between core/ and modules/, since Blink could not
include mixins across those layers.

Now that the IDL compiler supports cross-component mixins[1], the
workaround is no longer needed, and this patch replaces the partial
interface with interface mixins to follow the specs[2][3].

This CL has no behavior changes.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/6968299
[2] https://html.spec.whatwg.org/C/#the-localstorage-attribute
[3] https://html.spec.whatwg.org/C/#the-sessionstorage-attribute

Bug: 453570181
Change-Id: I2c8bcab18f22426cc61d8f25f94a8dd71e4cfb14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7074112
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Commit-Queue: Suyeon Ji <zeesuyeon@gmail.com>
Cr-Commit-Position: refs/heads/main@{#1534245}