Fix flaky timeouts in StructuredMetricsMixin::WaitUntilEventRecorded
This helper was using a fixed timeout of 3s, which is inadequate for
slow builds, such as sanitizer (win-asan) and debug builds. Switch to
using a fraction of TestTimeouts::action_timeout(), which is scaled
based on the build type. With this change, the test reliably passes for
me using the same config as win-asan.

Fixed: 450169384
Change-Id: Ie5ffe75169d2cafd0508b3d92d96db48ba1cdc09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7079029
Auto-Submit: Greg Thompson <grt@chromium.org>
Commit-Queue: Greg Thompson <grt@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534296}