[Android][Signin] Refactor SigninAndHistorySyncCoordinator.Result
SigninAndHistorySyncCoordinator.Result is refactored from an @IntDef
to a class that holds two booleans: mSignedIn and mHistorySyncOptIn.
This change is made to more clearly pass the results of the sign-in
and history sync flow to the SigninSnackbarController. The controller
needs to know exactly which actions the user performed during the
flow.

Note the particular edge case when ShouldSignOutOnDecline=True.
Upon declining history opt-in dialog HistorySyncMediator triggers
an async operation to signout the user. But we immediately call
BottomSheet's dismissHistorySync and then onFlowComplete and require
to know "did the user sign in during this flow". It is incorrect to
re-query IdentityManager immediately after a signout. Thus, I added
`didSignOut` in previous CL and use it in BottomSheetSignin and
FullscreenSignin.

Bug: 437039311
Change-Id: I4ff9dbb84db35f96bb22379443ad63cf28f7e0bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7062593
Reviewed-by: Menghan Yang <myuu@google.com>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Anton Goncharenko <agonch@google.com>
Cr-Commit-Position: refs/heads/main@{#1534390}