[Default Browser Promo] Handle uninitialized Feature Engagement Tracker
This CL addresses the edge case where the Feature Engagement Tracker
might not be initialized when the Default Browser Promo needs to
decide whether to display in the Magic Stack. This can occur, for
instance, during application relaunch.

To fix this, the promo's eligibility (from
Tracker.wouldTriggerHelpUi()) is now saved to ChromeSharedPreferences
when the activity is paused. So we can use this value when Feature
Engagement Tracker is not initialized.

Bug: 454385952
Change-Id: I430366d6bc5922810a46e9579592eab659ae85f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7032480
Reviewed-by: Xi Han <hanxi@chromium.org>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Xinyi Ji <xinyiji@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534448}