[reader mode] Fix HostZoomMapImpl_GetAllHostZoomLevels NOTREACHED crash
OverrideDefaultZoomForReaderModePage was calling a pattern that was
specified to be NOTREACHED within the GetAllHostZoomLevels code. It was
using SetZoomLevelForHostAndScheme to ensure that the RM zoom did not
apply multiplicatively with the default zoom. Modified to use
SetZoomLevelForHost which is fine since the host for RM pages is a
unique identifier and won't be shared with normal sites, so nothing is
lost by not specifying chrome-distiller scheme.

Bug: 454349127
Change-Id: Ib15368f8f49758e0c8b9d3b133d7a36a8c640d28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7076495
Reviewed-by: Brandon Wylie <wylieb@google.com>
Commit-Queue: Evan Luo <evanluo@google.com>
Cr-Commit-Position: refs/heads/main@{#1534455}