webview: migrate metrics to non-backed-up directory.
WebView's temporary metrics files were stored in the host app's data
directory, which will be backed up by many Android backup agent
implementations. This data isn't worth backing up and changes
frequently, so we want to avoid this.

Introduce a feature flag which defines a new metrics directory under the
Android "no-backup" files directory so that the backup agent knows to
ignore it, and will migrate existing files to the new location.

The migration is set up to be bidirectional: if a client previously had
the feature flag enabled but now does not, the files will be moved back,
to prevent metrics loss during experimentation.

Bug: 331195418
Change-Id: Icc57e0eacc4eae31696a4729295d003d926bf3c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7004435
Commit-Queue: Richard (Torne) Coles <torne@chromium.org>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534495}