Remove TrackedDeviceClass
Previously there was a `TrackedDeviceClass` interface used by the
XrBrowserTests; however, this type appears to be largely unused except
as a somewhat generic type (sometimes) for controllers. This simplifies
the existing usages down to just controllers and also reduces the max
number of devices/controllers to a more reasonable number (allowing a
disconnect/re-connect per test still). As a future cleanup, I think it
would be better to look at adding a secondary test interface and instead
of relying on the fake runtime to "pull" data, for the tests to "push"
data in a synchronous manner.

Also updates usages of `unsigned int` to `uint32_t`.

Change-Id: If55b3407dd8a68026ebe692868d004afb5f705c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7069794
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534499}