[GlicMI] Bind on pin (behind kGlicBindPinnedUnboundTab flag)
When a tab is shared in a conversation we check if that tab is NOT bound to any conversation and if that's the case, we bind it to the one it was shared in. Additionally we set the state so that when we navigate to that shared tab, SP is open.

Fixed: 454363384
Change-Id: I3f75bd34dc401d8576479c8df5385c9194ded34b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7076174
Reviewed-by: Bryant Chandler <bryantchandler@chromium.org>
Commit-Queue: Basia Zimirska <basiaz@google.com>
Reviewed-by: Marke Hallowell <wry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534504}