[AX] Fix spelling error bug with VoiceOver
This CL fixes a longstanding bug with VoiceOver not announcing spelling
errors.

The regression was introduced years ago by some performance work. We
reduced the amount of data serialized from blink, which included the
document markers stored on the inline text box nodes. Because the same
information was already available on the static text nodes, we removed
this duplication. However, we forgot to update the code in
AXPlatformNodeCoca::addTextAnnotationsIn that was reading the document
markers data from the inline text boxes. Our tests didn't catch that
regression, because they didn't have the tree structure we see in
production -- they were lacking inline text boxes under the static
boxes. We also fixed that in this change.

Finally, there was another, unrelated issue that caused caret navigation
to occasionally fail with VoiceOver. The NSRange used to mark a
misspelling in the attributed string could sometimes exceed the initial
NSRange provided by VoiceOver to Chromium. As a result, VoiceOver
remained silent instead of announcing the character. Although this issue
differs from the first bug, the two are closely related and should be
merged together to prevent regressions if either change lands
independently.

Testing: Tested with VoiceOver on mac and confirmed we get the same output as we do in Safari for spelling errors when we tab into the textfield, move by word, and move by character.

Fixed: 40649782
Change-Id: I4da01bb89df2c08c373d793a9d1691fe360976d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7080742
Auto-Submit: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1534530}