Add function for allowing populating Find Bar with selected text
This is part of a refactor to simplify how copy policy restrictions are
enforced in the find bar. Instead of hooking into clipboard and drag
events, the find bar will now directly query this function to determine
if it's allowed to use the selected text. This prevents potential data
exfiltration through the find bar when copy restrictions are in place.

Bug: 442804554
Change-Id: I132e2fcb66f7ceda8a041703fbad5662db65e7bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037215
Reviewed-by: Nancy Xiao <nancylanxiao@google.com>
Reviewed-by: Dominique Fauteux-Chapleau <domfc@chromium.org>
Commit-Queue: Aliaksei Bahdzevich <thelex@google.com>
Cr-Commit-Position: refs/heads/main@{#1534558}