media: Construct provisioning request with proper POST request
Instead of sending the provisioning request with the request in the URL,
we can use the request body to properly store and send the request.

This ensures that URLs do not exceed a certain length, and also makes
use of proper POST request techniques instead.

For safety reasons, this introduces a media switch which guards this
behavior, so we can turn it off and go back to the old behavior at any
time. This media switch is originally disabled by default.

We will roll this launch through experiments, so that it is safe and
does not break provisioning.

Tested by manually playing IC and verifying provisioning and playback
worked with no issues on Android.

Bug: 448700051
Change-Id: I2054017e67d034264fa366bc00cff76a7e960187
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7014207
Reviewed-by: Feras Aldahlawi <frs@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Vikram Pasupathy <vpasupathy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534560}