Use UserConfirmationDialog API for visiting sensitive origins
This CL has ExecutionEngine send a UserConfirmationDialogRequest
to the web client whenever it detects a navigation to a URL on
the Optimization Guide blocklist.

If the user accepts the dialog, the actor will be allowed to navigate
to that origin.

Already allowlisted origins will not have the blocklist applied to
them.

After this CL, ExecutionEngine::ShouldGateNavigation executes the
following steps:

1. Check origin allowlist synchronously, if origin is found continue
   the navigation.

2. Check the Optimization Guide blocklist. If the URL is on the list,
   defer the navigation and request a user confirmation dialog.

3. If the URL is not on the list or checking the list fails,
   defer the navigation and send a NavigationConfirmationRequest.

Bug: 439672418
Change-Id: If6482c3529d76ace09bec73a134b9cce4ddf7ce6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7075052
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Cr-Commit-Position: refs/heads/main@{#1534576}