Remove webgpucts module scheme from test targets
This module scheme is not ready for some services that read the test
results.

CL adds a copy of an existing webgpucts test suite (webgpu_cts_dedicated_worker_tests) that is ran only on the structured-test-id builder so that webgpucts results can still
be logged and checked for issues. The copy of the new test suite was tested on this try job:
https://ci.chromium.org/ui/p/chromium/builders/try/linux-structured-test-ids-rel-fyi/55/overview


Bug: 435467025
Change-Id: Ifb8daadfe66a48272171c12e7a5754a4915f8934
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7076183
Commit-Queue: Benjamin Joyce (Ben) <bjoyce@google.com>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1534597}