[GlicMI] Always close existing floaty when a new one is created
This CL ensures that if a floaty is shown by any mechanism,
we close any other floaties first. This is done by moving
the logic from `OnDetachRequested` to a new method
`OnFloatyWillShow` which is called from
`CreateActiveEmbedderForFloaty`. This bug was introduced
when we moved UnbindEmbedder and made it only be called
by tab binding. This new change is effectively the same
as what we did when we moved UnbindEmbedder. It ensures
that if we ever show a floaty for any reason, we make
sure all other floaties are closed.

Bug: b:453942053
Change-Id: Ic441794253ebfa647300d03200f0b4276a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7078991
Commit-Queue: Bryant Chandler <bryantchandler@chromium.org>
Reviewed-by: Trevor Perrier <perrier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534669}