net/disk_cache/sql: Pass entry key to SqlPersistentStore methods
This CL passes the CacheEntryKey to UpdateEntryLastUsedByResId and
GetEntryAvailableRange in SqlPersistentStore.

This is a preparatory step for a performance optimization where
operations will be distributed to different task runners based on the
hash of the entry key. This change is purely refactoring and has no
behavioral changes.

Bug: 443171275
Change-Id: I32b1a9971f3ce803039512317c29c41338fb5faf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7065293
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534683}