[TabIndicator] Correctly set bounds for actor indicator spinner
The actor indicator spinner is lazy-loaded in
MaybeLoadActorAccessingSpinner. So when OnBoundsChange for the
AlertIndicatorButton went through and tried to set the bounds for the
spinner, it would return early since the spinner hadn't been initialized
yet. This CL overrides the Layout(PassKey) view function so that once
the spinner is initialized, added as a child view, it triggers the new
Layout(PassKey) function and sets the correct bounds for the spinner.

This CL also moves the setting of the image size to the initialization
of the spinner since that would cause another Layout change causing an
infinite loop. Since the image size is static, we can just set the image
size at creation.

Screencast:
https://screencast.googleplex.com/cast/NTI0OTIwODkyNjg2MzM2MHxjODI5Y2QzMy1hMw

Bug: b:453726622
Change-Id: I96777207832a6d5a85e679cb48a96b728c03b934
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7080504
Commit-Queue: Jacob Stanley <jacobstanley@google.com>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Reviewed-by: Abe Boujane <boujane@google.com>
Cr-Commit-Position: refs/heads/main@{#1534721}