[soft navs] Ignore entropy check for video
We're finding that the entropy check fails on video sites sometimes,
even on videos which appear contentful. This can cause the seemingly
wrong element to be identified as the LCP element, i.e. the <video> is
ignored. This is likely a problem for hard LCP as well, although it's
potentially more noticeable for soft navs when clicking around a page,
e.g. on YouTube.

To work around this, this CL ignores the entropy check for video
elements for soft navs, and ignores it behind a flag for hard LCP. The
thought is videos are less likely to be used for gaming, and the status
quo does not work well enough for well-designed sites. We can revisit if
we find a better mechanism for video, or if we can consider more frames.

Bug: 434659232
Change-Id: I539ba23746162db4d47ebb18999b824d061f8f6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6981829
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534761}