Uprev ChromeScreenAI CIPD to 140.10
This CL updates the CIPD for downloading the library that is used for
ChromeScreenAI browser tests.

If browser tests fail, revert the rollout for the new binary in the
component updater as well. You will need to rollout a newer (or older)
release to update the clients that are affected by the faulty rollout.

Bug: 378472917

Change-Id: I3986368b59ea06cde5a3c40226cb989a39108d4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7063235
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534771}