Try skipping IPC channel pausing when launching renderer process
We pause the mojo channel to the renderer when launching the process,
until we get the OnProcessLaunched signal, which is actually posted
early in the launcher thread but is delayed by the busy main thread.
This is causing any message sent before the OnProcessLaunched signal
is run on the main thread to be delayed, potentially for quite a
while, especially at browser startup. We want to experiment with not
doing the pausing here, and instead send the init messages early on
to ensure they are run before other messages are sent to the
renderer.

This CL introduces kSkipIPCChannelPausingForNonGuests, which will
skip the IPC pausing when enabled, except for GuestView processes,
which has a dependency on speculative RFH creation, making it
dependent on IPC pausing.

This CL also moves the timing of firing OnRenderProcessHostCreated
to RenderProcessHostCreationObservers to be done earlier, firing
when the renderer process launch is requested instead of when it
finishes launching. This is so that important initialization messages
triggered from the observers can be sent as soon as possible, before
other messages like navigation commit (which can now happen earlier
because there's no channel pausing).

However, some previous OnRenderProcessHostCreated listeners actually
want to wait until the renderer process is launched (so that they
can get the PID, log launch metrics, etc). This CL introduces a new
OnRenderProcessHostLaunched to RenderProcessHostCreationObserver,
which is fired when the renderer process already launched (which was
the timing of OnRenderProcessHostCreated before), and moves some
of the previous listeners to this new function instead.

Bug: 448511116
Change-Id: I7e33e6cef48392ccf794b8f159b3df14dd485caf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7031539
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534779}