ash: Reduce tablet mode's multitask menu hit region
In ChromeOS tablet mode, with WebUI tabstrip disabled (soon to be the
default), the multitask menu in the top center of a window makes it hard
to initiate drag of a tab located in that area. Drastically reduce the
height of the menu hit region to mitigate this. Even then it's still
reliably easy to trigger the menu by swiping from top. The one exception
is the lower window in a horizontal split (portrait mode), for which
we'll find a better solution.

Bug: 421465978
Change-Id: I9a9278b860897ea34aca2d1293e213422cbbb2bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7065652
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Georg Neis <neis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534781}