[//gpu] Remove OOP_RASTERIZATION
The only meaningful thing that this is used for is RawDraw, which
verifies that the client has set it along with
SHARED_IMAGE_USAGE_RAW_DRAW. The only client that sets RAW_DRAW usage
is GpuRasterBufferProvider [1], which by definition always uses
accelerated raster (which is of course now always OOP-R). Hence, it's
not necessary to preserve this check on the service side.

[1] https://source.chromium.org/chromium/chromium/src/+/main:cc/raster/gpu_raster_buffer_provider.cc;l=298;drc=c99aa55ee638df4d6f0073c5d950acbda6ab4c6d;bpv=1;bpt=1

Bug: 391648152
Change-Id: Idcfedfb9457244e9a05b2d4844703359387c8a20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7077821
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534788}