Change the WebView target used for orderfile generation
Change the WebView target used for orderfile generation to target more
devices. Indeed, the system_webview_64_32_apk is the config we ship to
more than 80% of devices. Also, when we are generating a 64-bit
orderfile, we should run the 64-bit library. For
system_webview_32_64_apk, the renderer run in 32-bit mode. So it is not
actually using the 64-bit library.

Bug: 435633362
Change-Id: Ic6845b6917a4bb9ea30dccd0aba762ac97d83b32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7072791
Commit-Queue: Trung Nguyen <tne@google.com>
Reviewed-by: Haiyang Pan <hypan@google.com>
Cr-Commit-Position: refs/heads/main@{#1534839}