Fix specificity of declarations in @contents.
Assume we have declarations like this:

  @mixin --m1(@contents) {
    color: red;
    @contents;
  }
  div {
    @apply --m1 { background-color: green; }
  }

Internally, after cloning and such, we get roughly this object tree:

  div [StyleRule] {
    MixinRule {
      & [outer fake StyleRule] {
        StyleRuleNestedDeclarations {
          & [inner fake StyleRule] {
            color: red;
          }
        }
      }
      StyleRuleContents {
        & [outer fake StyleRule] {
          StyleRuleNestedDeclarations {
            & [inner fake StyleRule] {
              background-color: green;
            }
          }
        }
      }
    }
  }

During parsing, StyleRuleNestedDeclarations copies the selector from
its parent rule, which during mixin parsing is a dummy & rule.
During cloning (for mixin application), we reparent these rules;
the parent pointer in the StyleRules is updated, and the selector
in the StyleRuleNestedDeclarations is copied from its parent.

However, StyleRuleNestedDeclarations does not actually want & as its
parent; it wants div. (This matters especially if the outermost rule
has a selector list, with differing selectivities.) Somehow this worked
for the normal @apply case, but not for @contents. The reason is that
StyleRuleMixin does not actually hold the fake parent rule itself;
it is a StyleRuleGroup and holds the child rules directly. So the fake
parent rule, while it _is_ the parent rule of the rules in the @mixin,
does not come into play; when cloning, we reparent the rules within
to have the mixin's parent (which is div).

We need to do the same for @apply; it is _not_ a StyleRuleGroup
but should perhaps be so (the spec writer is not entirely decided),
so that instead of cloning (outer) fake StyleRule holding its
@contents block, we should skip over that and clone/reparent
the rules within directly. (We used to simply not clone @apply
and @contents rules in StyleRule::Clone(), since the caller would
clone their fake parent rule itself. But as we've seen, cloning
the fake parent rule means that the selector is not replaced,
only reparented.)

Fallback blocks in @contents had the same issue; added a test
that is similar to the existing ones.

Bug: 406935599
Change-Id: I328a9a8026d82c23b7cc4543f066baf2f83a2e6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7080794
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534958}