[Browser Layout] Move params to a new file with utility methods
This moves BrowserLayoutParams to its own file and adds a method to
get a params block relative to a particular rectangle. This can allow
us to translate to local coordinates when laying out the browser view
or the top container.

It also adds a method to BrowserViewLayoutDelegate to return a set of
layout params in the BrowserView's local coordinates, to make
calculating the new layout easier.

Tests for the new functionality have been added.

Bug: 453717426
Change-Id: Ida8458989bd396b7b9d8a3b383f37961fc4a80c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7081751
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534963}