[Browser Layout] Push more code down into the old implementation
This moves logic for calculating dialog bounds that was in the base
BrowserViewLayout class into BrowserViewLayoutImplOld, with a few
virtual methods that will need to be implemented by the new layout.
This removes all of the dialog layout code specific to the old layout
from the base class, meaning the new layout can do whatever it wants.

A test-only method that is too implementation-specific is also tagged
at declaration and point of use for later removal.

Bug: 453717426
Change-Id: If8cdbe1e2fa4f377ab2a198829daa5b553e03b27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7081752
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534964}