[clank][geo] Add OS-location warning to location settings subpage
This CL touches the location settings subpage that allows, when the
kApproximateGeolocationPermission feature flag is enabled, to switch
between approximate and precise location grant for a site. In
particular it adds a warning in case the granted accuracy (precise) is
higher than the accuracy granted at the OS level (approximate). The
warning appears and disappears depending on the user selection.

Bug: 437026554
Change-Id: Iacd9db1df130ea1b956d89dc92bb60348fe0f01f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7072231
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535014}