Remove superfluous proto/optimization_guide includes from Lens headers
Remove protobuf and optimization_guide includes that either are not used
or can switch to forward declarations. Then do IWYU to fix the build.
Fix a few nits with member variables along the way.

The AI sort of helped by trying to identify what part of a given include
is being used in a Lens header.

Bug: 40318405
Change-Id: I37eeb6c3defeab1c9be5dc20004d0f713cb04dc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7055329
Reviewed-by: Duncan Mercer <mercerd@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535172}