[Wayland] Stop treating tiled window state as maximized
Normally, the tiled window state reported by the Wayland server
indicates which edge of the screen a window is tiled in (eg. a window
on the right side of the screen will be tiled right, a window in the top
left corner will be tiled top and tiled left). However, some compositors
just set all edges as tiled for any tiled window. Normally this should
only happen if a window is maximized, and the expectation is that we
would use the server's maximized state to decide how to draw the window,
however there are some compositors (eg default GNOME in Ubuntu 24.04,
but for some reason not GNOME in Fedora) that set all tiled edges for
any tiled window. This CL removes the maximized check so windows draw
properly (without rounded corners) in this case.

R=thestig

Change-Id: Ib8e109544c1a9763bbecea024ad59ed1089b8c51
Fixed: 454083284
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7083267
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535196}