Fix and reenable PreinstalledWebAppsBrowserTest.CheckInstalledFields
I have no idea if this test is still flaky on Chrome OS or not, but
having it disabled means it bitrotted a lot. Update the test to match
the current set of default apps and re-enable it to find out if it
is still flaky.

Also adds missing only_use_app_info_factory to the google chat app
configuration, to make the test pass.

Bug: 328691719
Change-Id: I47e96ae8e13145e0e5b4109a0017d65c57cda24e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7081018
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535205}