TaskAttribution: Add BlinkTaskState trace events for microtasks
This CL primarily adds trace events with task state information for
microtasks, so we can get the task_state_id. We already output this for
async task propagation when creating a TaskScope, e.g. postMessage,
setTimeout propagation, but don't for microtasks since this is handled
by v8. This will be valuable to diagnose soft nav detection issues
related to task attribution.

To add these trace events, we use v8's PromiseHook API to get callbacks
before and after running Promise-related microtask, as well as add hooks
into EventLoop for internal microtasks. Unconditionally registering a
PromiseHook would likely be too expensive, so we mitigate this by only
registering the hook when tracing is enabled for the relevant category,
and we only check this at construction time and when tracing is turned
on (via AsyncEnabledStateObserver).

This also unifies the existing trace event, renaming BlinkTaskScope to
BlinkTaskState, and moving this to a new "task_attribution" trace
category (given the additional PromiseHook overhead and main use case, I
wanted to separate this from "scheduler", which has other use cases).

Bug: 447555069
Enabled-by-default-reason: killswitch
Change-Id: I1579ca66c60202ed4f597d7167ec37df3b5aff7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6986393
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535210}