[Read aloud] Install enabled langs when the tts engine is ready
First time users seem to be running into a race condition where
onVoicesChanged gets called before the extension is ready on the Chrome
side, so we never try requesting any installs.

This change now also requests a language install as soon as the engine
extension is ready, so if this happens we still request installation. If
onVoicesChanged was already called and the initial installation already
happened, we'll receive that status and do nothing.

b: 441587517
Change-Id: I61cbcfbcb92768887e8aea04915cc278cc3b373e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7076154
Reviewed-by: Eitan Goldberger <eitang@google.com>
Commit-Queue: Kristi Saney <kristislee@google.com>
Auto-Submit: Kristi Saney <kristislee@google.com>
Cr-Commit-Position: refs/heads/main@{#1535217}