[Chrome Next] Start of new layout supporting toolbar height side panel
This implements all of the API around the layout, including a data
structure to hold the layout, but not the layout itself, which will be
added in a follow-up.

This also adds a flag that switches the implementation over to the new
layout for "normal" (i.e. tabbed, non-app) browsers only. The layout
does not yet work, of course.

Bug: 453717426
Change-Id: Id77214e5075aa05a49781830d95996ab5f4d2e60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7082967
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535229}