[TabAlert] Properly show media capture alerts
The media recording alert was incorrectly shown when only the audio or
video is being recorded for a tab. This CL fixes this bug by having
the alert controller check which media alerts are currently active
each time a media alert state changes.

The desktop capturing alert should also be active when window capture
or monitor capture is happening. Previously, it was possible for the
desktop capturing alert to become inactive if a tab is being captured
by the window and display at the same time and one of the
capture sources stops capturing.

Fixed: 454088668
Change-Id: I76dfd4868cae5ddf140cc007d895e75e752bda33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7080494
Commit-Queue: Steven Luong <stluong@chromium.org>
Reviewed-by: Alison Gale <agale@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535322}