[PDF Ink Signatures] Fix caret appearing when form field is focused pt.1
When a form field has focus, and the user moves focus to a different
form field using Tab/Shift+Tab, the PDF caret appears when it should
not. This is because PDFiumEngine::OnFocusedAnnotationUpdated() calls
PDFiumEngine::ClearTextSelection(), which always sets the caret to be
visible.

Revert the changes done to ClearTextSelection() in
https://crrev.com/1525714 and instead have PdfCaret handle setting the
visibility when it calls it. Other existing ClearTextSelection() callers
will need to be reviewed on whether they need to set caret visibility or
not. For now, assume they do not need to.

Bug: 454069614
Change-Id: Ib6be43249b46bee766e2d4e6f3cc8df855885983
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7083308
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535326}