Moved UMA TabGroups_SavedTabGroups recording to SavedTabGroupUtils
We put all the UMA recording for metrics starting with
TabGroups_SavedTabGroups to a single class. These items record when
the user clicks one of the first four items in the tab group submenu.
We also add a new metric so that the action to move a tab group to a
new window is recorded.

Bug: 454689946
Change-Id: Ieb8b90ada2bc11e67fccc3df9323bb9c2a1f12a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7081522
Reviewed-by: David Pennington <dpenning@chromium.org>
Reviewed-by: Alison Gale <agale@chromium.org>
Commit-Queue: Dominic Austria <dominicaustria@google.com>
Cr-Commit-Position: refs/heads/main@{#1535332}