[GlicMI] Expose Glic window activation callback.
In multi-instance, the focused browser manager isn't able to register
its own callback for changes to glic window activation state, so we
expose this via the delegating sharing manager that the instance uses so
it can be called directly. We hold the state in the delegating sharing
manager in case changes happened while delegating to another underlying
sharing manager.

Bug: 452150693
Change-Id: I1a8ddfe743d84ee2e4574a50990b7484f4ac9756
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7081116
Reviewed-by: Bryant Chandler <bryantchandler@chromium.org>
Commit-Queue: Marke Hallowell <wry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535378}