[Reading Mode]: Add Immersive Reading Mode overlay surface and view
This change introduces the foundational view for the new Immersive
Reading Mode feature. It creates the ImmersiveReadAnythingOverlayView
which will act as the host for the reading mode WebUI when in its
full-page state.

Following the design document (go/immersive-readingmode-designdoc), this
new view is added as a child of the ContentsContainerView and managed by
the ContentsLayoutManager. This ensures it is correctly stacked above
the web contents and will work properly with split-screen views.

A new browser test is added to verify that the overlay view is correctly
added to the view hierarchy with the expected default properties.

Bug: 449167659
Change-Id: If4cf11ebee8b876026db6050285b9257136fc475
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7016652
Reviewed-by: Kristi Saney <kristislee@google.com>
Commit-Queue: Jacobo Misael Tapia <jatapiaro@google.com>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535465}