window management: display category only if topology is enabled
Window Management API is enabled only if Window Management feature flag
is enabled and display topology is available on this device [1]. This
category should also be visible with these requirements are met.

If the user has already granted permission to a website, the user will
not be able to revoke it from this category if device doesn't have
display topology (because the category is invisible). This is
acceptable, because if the device doesn't support display topology, then
Window Management Api isn't available, and this permission has no
impact.

1. https://source.chromium.org/chromium/chromium/src/+/main:components/permissions/contexts/window_management_permission_context.cc;l=47

Bug: 450799993
Test: local run
Change-Id: I629f973c71f988948dc33a2be739a948b50e826c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7079124
Commit-Queue: Daniil Hrybchuk <renkens@google.com>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535690}