Automatically show Customize Chrome IPH on NTP load.
Adding IPH feature flag, its registration and logic for auto show on NTP
load. Right now both IPH and Customize Chrome are tried to be opened at
the same time, the relation between panel opening and IPH will be
established in next CLs. The feature is now started after
CustomizeChrome button on NTP is shown (this way we're sure we can try
to open IPH). Because of that, the tests had to be reorganized too.
Extracted logic from browser tests to test base. As IPHs don't have to
be checked in every test, new test class specifically for them was
introduced. Other browser tests for this feature now are interactive too
(to be able to wait for the button to show).

Bug: 443741393
Change-Id: I26fd00d8a316ff2aab29b715709ea01d91656175
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7060813
Reviewed-by: Ryan Sultanem <rsult@google.com>
Commit-Queue: Kacper Bajkiewicz <kbajkiewicz@google.com>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535886}