[Pref Tampering] Persist deferred validation reset list to PrefService
1. For the async validation pass, the DeferredEncryptorRevalidation was
correctly adding resets to its cloned kTrackedPreferencesReset list, but
this list was never propagated back to the live PrefService, so the
record of the reset was lost. This change updates live PrefService list
with the deferred pass's list to ensure it's persisted.

2. Due to the async nature of the PrefService initialization, the
kTrackedPreferencesReset list turned out to be empty when queried by the
UI, when resets occurred during the sync FinalizeFilterOnLoad pass. This
CL post a task to directly update the list after the PrefService is
successfully created.

Bug: 454384390
Change-Id: I4e10ca4cf3e3b8ec209eead711969c4844c2ba34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7075053
Reviewed-by: Anunoy Ghosh <anunoy@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Zack Han <zackhan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535899}