[clank][permissions] Fix string on multiple os-level missing permissions
The string in PageInfo saying "Turn on permissions for Chromium in
Android Settings" when multiple permissions are missing at the OS
level was not being displayed because of a bug that I introduced with
https://chromium-review.googlesource.com/c/chromium/src/+/7057282.
This CL fixes that.

This would need tests. Unfortunately it's not immediate to add them
because we need to mock the OS-level permission state. I'll follow up
with that with another CL.

Bug: 437026554
Change-Id: I3d208203a1cb34fc4cb56dbb610a3ba66b9ca34f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7084499
Auto-Submit: Antonio Sartori <antoniosartori@chromium.org>
Commit-Queue: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1535976}