[WebView] Use a unified platform string for reduced User-Agent
This change modifies the GetUserAgent() function to return a hardcoded,
unified platform string when user-agent reduction is enabled for
WebView.

This change enforces the platform string to be "Linux; Android 10; K;
wv" for all reduced WebView User-Agents, regardless of the underlying
form factor. This ensures a consistent UA string that aligns with the
CTS test expectations.

Change-Id: Ib773388108dc916567c91ef10a88cfc6dd5f7cfb
Bug: 453862825
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7081663
Commit-Queue: Victor Tan <victortan@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536034}