Add CreateBoundSession to DeviceBoundSessionManager
This CL creates a new method CreateBoundSession on the
DeviceBoundSessionManager. This will allow the browser process to create
a new Device Bound Session in the network process and set associated
cookies.

This is needed to support OAuthMultilogin, so that users signed in to
Chrome can safely restore cookies without introducing a gap in DBSC
protection.

NO_IFTTT=No change to SessionParams fields

Compile-Size: Size increase is in parent change.
Bug: 439821296
Change-Id: I2a9715153859c79332f375b2e7f67a7c2de046fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7040428
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Fred Shih <ffred@chromium.org>
Reviewed-by: thefrog <thefrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536203}