[iOS] Add and Implement Location Bar Badge Mutator
Created a mutator to handle a user tapping the Location Bar Badge. This
CL builds the foundation for the mediator (which is the mutator) to send
relay calls to show UI. Also moves badge_type.h to model/ in order to
avoid dependency cycles and to abide by folder structure best practices.

Fixed: 454070017
Change-Id: I2f28bcf747acd1f1b500c4a88da879dc26e335ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7075732
Reviewed-by: Adam Arcaro <adamta@google.com>
Commit-Queue: Joemer Ramos <joemerramos@google.com>
Cr-Commit-Position: refs/heads/main@{#1536219}