actor: Remove redundant validation from select tool
The select tool performs a synchronous check after changing the value
that the current value is set to the new option.

In practice, sites commonly use JS to asynchronously mutate the values
of a <select> so this check can sometimes cause a false negative, the
value does settle to the expected option but the check returns failure.

Since the actor is able to tell what the new value is through a new
observation this check is redundant anyway so removing it will avoid
this failure case.

Bug: b:453529066
Change-Id: Ib4983549aaacf74ac123a2e984ef5caf4349f529
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7088307
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Auto-Submit: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536228}