actor: Use more precise terminology in ActorTask
Change some of the terminology used in actor task to be more precise /
less easy to misinterpret.

In particular "active" is misleading since tasks may be running but
paused or waiting for user input; and we put tabs into "active" state
but a tab being "active" also means it's the selected tab in the tab
strip.

Change the terminology to be that a task can be either:

ActorControlled - the task is Created, Acting, or Reflecting. i.e. in
this state the user cannot interact with the tab and the input-blocking
scrim is shown.

UserControlled - the task is still in progress but has yielded to the
user, either through actor or user request. The scrim is removed and the
user can interact with the task's tabs but the task can resume to actor
control.

Completed - the task is no longer running. It either finished
successfully or was terminated.

Change-Id: I3175014cbc4bb9cbf0ce496046cc21873fb988f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7082970
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Auto-Submit: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536235}