Fix restoring window on Wayland
On Wayland, if a window was maximized and then minimized, calling
Restore() on the window will first trigger switching to normal mode, and
then after a moment to maximized mode. This commit changes the behaviour
so it go straight into maximized mode without in-between states.

Bug: 40946306
Change-Id: I372d992ff17b977e84a28f828f6024f8c2f774a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7079799
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Reviewed-by: Kramer Ge <fangzhoug@chromium.org>
Commit-Queue: Patryk Chodur <pchodur@google.com>
Cr-Commit-Position: refs/heads/main@{#1536259}