mac: Prevent reentrancy in ImmersiveModeFocusSearchMac
ImmersiveModeFocusSearchMac connects BrowserView, the overlay widget,
and the tab overlay widget for focus traversal. When FocusManager cannot
find the next focusable view in the current widget, it queries this
parent traversable.

If the search cycles back to a widget that also fails to find a view, it
re-enters ImmersiveModeFocusSearchMac, causing an infinite recursion and
stack overflow. This is actually happening in the wild with a low crash
rate.

This CL adds a reentrancy guard to FindNextFocusableView to break this
cycle by returning nullptr if a search is already in progress.

Bug: 455109313
Change-Id: I2e086aa263834c4e97d2f7b441a0a9e48c37a37f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7085119
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536275}