Get rid of backup keyset cleanup result metric
This metric is related to the VK -> USS migration, where we've already
gotten rid of most of the other relevant metrics as obsolete.

We can also get rid of the cryptohome code that reports these but that
will of course be a separate change.

Bug: b:455573243
Change-Id: Ic43f2e330d33b6e7ff06a9b9dfdd55f89c798a2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7087827
Commit-Queue: Zi Lin <lziest@google.com>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536278}