[rustmojo] Restore some old rust bindings gn code
We're starting to get to the point where we actually need to invoke the
generator, so this CL restores a subset of the old rust code for doing
so. It stubs things out so we can grow them in the future.

The exception is the gni code, which was long and complex. To make
things more maintainable, we move the old code to its own file which is
called from the main gni file. We preserve all of it under the
assumption that we're going to need it soon, e.g. wrangling
dependencies.

Bug: 444509367
Change-Id: I74dc117d94c2496fb92fd0e28ae6967116294bc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7082227
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1536282}