[Contextual Tasks] Create UI service for incognito and guest
Since the ContextualTasksContextController now supports guest and
incognito profiles (https://crrev.com/c/7081083), we can now support
them from the UI service as well. This patch updates the factory to
create the ContextualTasksUiService for guest and incognito profiles.

This also addresses a crash when the feature flag is enabled and the
user attempts to open an incognito tab. The change to IsAiUrl is to
address what is most likely a red herring crash stack caused by the
same flow.

Bug: 452127302, 455108064
Change-Id: Ifdb877aeada0d75d239e1dbfb04097d7836d3e55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7087460
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Mickey Burks <mickeyburks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536323}