WebUI: Disallow @ts-expect-error, @ts-no-check via ESLint checks.
This is to more accurately enforce the styleguide at [1].

As part of this CL fixing a few remaining violations in the codebase, by
modifying the code. In a couple of cases in the CrOS Settings codebase,
suppressing the ESLint checks since the violations are not trivial to
address.

[1] https://google.github.io/styleguide/tsguide.html#ts-ignore

Fixed: 454067870
Bug: b:266455078
Change-Id: I8e98ed9166102f82774f668b6ed234fba15c415a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7081017
Auto-Submit: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536356}