extensions: Remove context menu callbacks from EAVC
This is a part of the patch series to move stateful UI logic from
ExtensionActionViewController to ExtensionActionPlatformDelegate.

ExtensionsToolbarContainer needs to observe shown/closed events for
context menus attached to toolbar action buttons to update some internal
states. And we used to notify these events in the following path:

  ExtensionContextMenuController::ShowContextMenuForViewImpl()
  -> ToolbarActionViewController::OnContextMenuShown()
   = ExtensionActionViewController::OnContextMenuShown()
  -> ExtensionsContainer::OnContextMenuShownFromToolbar()
   = ExtensionsToolbarContainer::OnContextMenuShownFromToolbar()

To reduce UI logic from ExtensionActionViewController, we don't want
ExtensionActionViewController to appear in the middle.

To achieve this, this patch reworks how these events are notified. We
introduce the observer interface for ExtensionContextMenuController that
allows its owner to subscribe to context menu events. The observer
interface is implemented by two classes with following behavior:

  1. ToolbarActionView: notifies its owner via a delegate.
  2. ExtensionMenuItemView: does nothing.

ToolbarActionView::Delegate is already implemented by
ExtensionsToolbarContainer, so it can detect context menu events.

The path for the example above will become:

  ExtensionContextMenuController::ShowContextMenuForViewImpl()
  -> ExtensionContextMenuController::Observer::OnContextMenuShown()
   = ToolbarActionView::OnContextMenuShown()
  -> ToolbarActionView::Delegate::OnContextMenuShown()
   = ExtensionsToolbarContainer::OnContextMenuShown()

This is a pure refactor. No functional changes are expected.

Bug: 448199168
Change-Id: I01014d8cfafaf6479e518ec5b82ab29d23cf0d7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7052664
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536386}