Use mime type for context node image requests
RequestImageForContextNode previously returned a file extension, which
could be inaccurate if the image was encoded into a different format.

This change modifies the function to return the correct MIME type of the
resulting image data. Callers have been updated to use the new mime_type
field.

To maintain compatibility with the Java side, the Android implementation
derives the file extension from the mime type until the Java code is
updated.

Bug: 448897265
Change-Id: I28788664176399deef5d6eb160e3095ae6ecb561
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7083655
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Commit-Queue: Gary Klassen <gklassen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536407}