Add a V4 to the DomainDiversity metrics
While a computer is sleeping, TimeTicks do not increase. This means that
a "24 hour" PostDelayedTask will run every 24 hours of unsuspended time,
which is not what is intended here.

This CL introduces a V4 of the metrics which runs on its own
WallClockTimer instead of a PostDelayedTask.

Note: This is an additional startup task. If folks see some additional delay in startup shortly after this lands this might be the culprit.

Bug: 453632345
Change-Id: Ic712edd7fb4f55664c3cef54e80df4b868aef14c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7073111
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Josh Karlin <jkarlin@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536603}