[ios] Prevent bookmarks' new folder's edit flow gets interrupted.
This is a real issue on prod, reproduce steps: Have some account signed in with sync enabled; Tap the create folder button; Editing status of the cell gets interrupted after a few seconds while idle or typing the new name.

A few EGTests were flaky because of the editing flow gets interrupted.

Another issue fixed in util_testNewFolderNameCommittedOnNavigatingAway is: the cell attempted to tap can be partially hidden because of iOS 26 layout changes.

For the remaining disabled test cases, they seem all to be flaky, so use the FLAKY_ instead of DISABLED_ to gather more logs.

Fixed: 440485616, 440475041, 440572503

Change-Id: I007c5affd972e79dab52305757ba09e10e09b099
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7081653
Commit-Queue: Huiting Yu <huitingyu@google.com>
Reviewed-by: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536735}