Remove redundant map lookups in btm_service_impl.h.
Instead of doing a `contains(key)` followed by `.at(key)`, reuse the
iterator from the initial container check. This avoids the duplicate
lookup and need to re-calculate the hash from the string in the key.
Also, rely on the bracket operator's behavior of default-constructing
the value if the key doesn't exist, in order to avoid a redundant map
lookup.

Bug: 452950815
Change-Id: I6a00bb325d97f9d2367827c16004f679fc57a7d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7087488
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Commit-Queue: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536757}