[a11y] Return empty role description for `kSvgRoot` role on Android.
Talkback does not use the role description for SVGs. If they have an
aria label and/or a SVG title tag, they are announced as:
  "[Aria label], Image, [SVG Title Tag]"
and when they don't have either of the above, they are considered
uninteresting and not added in a11y navigation.

Return an empty role description for `kSvgRoot` and update tests.

AX-Relnotes: n/a
Bug: 447360631
Test: Locally on Talkback 16.2
Change-Id: I915c6623c9a5a35e9bfa977c87fdafb0deb05419
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7088748
Reviewed-by: David Tseng <dtseng@chromium.org>
Auto-Submit: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536772}