Make ChromeAndroidTaskTracker.createPendingTask return a nullable value
Currently an exception will be thrown if a pending ChromeAndroidTask
can't be created. A better way is to return a null ChromeAndroidTask so
that callers (such as native code for extensions) can handle the case
properly.

Additionally, when ChromeAndroidTaskTracker.createPendingTask() should
return null, we will also invoke its callback with value 0 (nullptr) for
consistency.

Bug: 453777179
Change-Id: Ia2cb66d416b6c5a0426e69da7666abeb31ea48c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7087572
Commit-Queue: Linyu He <linyuh@google.com>
Reviewed-by: Jack Thiesen <jthiesen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536779}