[Vertical Tabs] Fixing Flaky Test for Tab Search Button
This CL offers a potential fix to a flaky test on linux
(VerifyTabSearchVerticalToHorizontal). Since the tab search button is
now unpinned when we enter the vertical tabs mode, sending the escape
character is no longer necessary to close the bubble.

Additionally, I consolidated the Test Case Blocks using the Do()
function. This will allow for better readability and understanding if
failures occur because we will easily know which step failed.

Bug: 455030583
Change-Id: I939f426e0378914d53e9100116b4aa5bc066f70e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7087989
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Kunal Daftari <kunaldaftari@google.com>
Reviewed-by: Shibalik Mohapatra <shibalik@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536785}