Fix restoring maximized window on X11
On X11, if a window was maximized, set fullscreen and then restored
using the Restore() function, the window end up in a weird state in
which it is not maximized but still in the fullscreen mode. This commit
fixes that so it just goes from fullscreen to maximized. Calling
Restore() once again causes the window to switch to normal mode.

Bug: 40946306
Change-Id: Iff0e00181b251efb96caea41d3fefb7d9b169474
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7088752
Reviewed-by: Robert Ferens <rferens@google.com>
Reviewed-by: Kramer Ge <fangzhoug@chromium.org>
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Commit-Queue: Patryk Chodur <pchodur@google.com>
Cr-Commit-Position: refs/heads/main@{#1536804}