Glic Entrypoint Variations: only use kColorSysOnSurface when using the alt background color
We were using ui::kColorSysOnSurface for the foreground color even when
the flag was off. Some custom themes from extensions can make the
surface colors dark while Chrome thinks it's a light theme, so
kColorSysOnSurface is dark and can't be seen against the background.

This CL changes back to using a foreground color ID that works in this
situation and is drawn as white.

However, we'll still need kColorSysOnSurface when the alt icon is
enabled because we set a background color that matches Chrome's
light/dark theme rather than the custom theme.

Bug: 455014439
Change-Id: I005154b1d958a4ce607d9bfb60cd0496905689a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7087595
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Ian Wells <iwells@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536836}