Improve download failure logging in InstallerDownloaderModelImpl
This change refactors how download failures are logged in
`InstallerDownloaderModelImpl`. It introduces a helper function
`RecordDownloadInterruptReason` to centralize the logging of download
interrupt reasons. The `INTERRUPTED` and `CANCELLED` download states are
now handled separately, ensuring that the specific
`DownloadInterruptReason` is recorded only when a download is
interrupted, not when it's cancelled.

Bug: None
Change-Id: Ia225405dabdf8615847e97449ec46c9d6493030e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7092493
Reviewed-by: Kaan Alsan <alsan@chromium.org>
Commit-Queue: Daniel Soromou <koretadaniel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536860}