[AIM] Fix carousel and submit being clickable when hidden.
In the case of the submit button, the `div` is marked disabled instead
of the button which leaves it clickable. This change adds a guard in
`submitQuery_` that causes it to be a no-op when submit is set to
disabled.

In the case of the carousel, conflicting changes were merged resulting
in the wrong `part` attribute being exported. This change corrects the
CSS to use the correct one.

Bug: 455872614
Change-Id: I2f2d3d2609d9ea4bb7467acfc4df5c6619e5dad4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7091238
Commit-Queue: Juan Mojica <juanmojica@google.com>
Reviewed-by: Marlon Facey <mfacey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536864}