Apply optimization guide checks when origin gating enabled in MayActOnTab
This CL fixes a bug in navigation gating where the optimization
guide blocklist may not be applied if the browser starts the task
on a sensitive site. This CL applies changes navigation gating to
keep using the optimization guide blocklist during safety checks
before every action.

The CL ensures that URLs whose origin has been added to the per-task
allow list do not have the optimization guide check applied. This is
in case that a user does confirm in the dialog to interact with the
sensitive site.

Bug: 439672418
Change-Id: Ib8b711899177fc0b353a2fabe7d134603191580c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7088088
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536911}