PaintTiming: Fix initialization of TextRecord frame index
There's a slight difference between how the frame index for TextRecord
and ImageRecord is initialized: TextPaintTimingDetector passes the value
to TextRecord in the constructor (except for largest-ignored), whereas
ImagePaintTimingDetector sets it for ImageRecords when queueing the
record for paint time. This worked for TextRecords because they are
created in the same frame as presentation feedback is desired (except
for largest-ignored). For images, the records are typically created
before they're ready for presentation time, since the detector needs to
wait for the image to be (sufficiently) loaded.

Unfortunately, the refactoring in crrev.com/c/6781530 broke the
TextRecord version by not setting the value in the constructor. This
can lead to earlier than desired paint times for records, depending
on timing.

This CL fixes this by setting the frame index on record when queueing it
for paint time, which is the pattern used for ImageRecords. This also
adds a unit test to test that we don't dequeue records too early.

Fixed: 455791378
Change-Id: I4a9cf964d5dd27feffa34311d28f1273556cc01d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7091357
Reviewed-by: Johannes Henkel <johannes@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536915}