Ensure parity for Opus and opus in codec strings
MediaSource.isTypeSupported returns true for Opus support, but in
SourceBuffers "Opus" with a capital O does not return support.

A similar condition happened and was fixed for flac, where fLaC support
was inconsistent, and it was fixed in crbug.com/40856553.

The bug also contains more information about diff browser support info
and a detailed rationale and explanation.

test: added unittests.

Bug: 40897078
Change-Id: Ia8e205d44a6823f368402a72ff5197a8e763fc58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7085616
Commit-Queue: Vikram Pasupathy <vpasupathy@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1536951}