Fix the default return value for ShouldPauseChannelUntilProcessLaunched
This function should return true when the flag
kSkipIPCChannelPausingForNonGuests is disabled (because we should pause)
and false when the flag is enabled (because we should not pause).

This function previously returned the opposite value from what we want,
and this CL fixes that and adds comments to avoid confusion.

Bug: 448511116
Change-Id: Ia9e61f8e8ad4a59e65f215dde1621fb6f97197c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7094238
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Auto-Submit: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537164}