Adjust Finch experiment for bookmark count limit to be more targeted
This CL refines the Finch experiment for the bookmark sync limit.
Previously, the experiment would assign a new limit to all users in the
experiment group, regardless of their current number of bookmarks.

This change adjusts the logic so that the experimental limit is only
applied to users who have already exceeded the default limit of 100,000
bookmarks. This ensures that the experiment only targets the users who
are actually affected by the problem of silent sync failures due to the
hardcoded limit, making the experimental results more accurate and
relevant.

The default limit for users not in the experiment or for users in the
experiment with fewer than 100,000 bookmarks remains unchanged.

Bug: 454302754
Change-Id: I85bf5162622c73aadf67cf45341b85b68736dedb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7089524
Commit-Queue: Ankush Singh <ankushkush@google.com>
Reviewed-by: Ankush Singh <ankushkush@google.com>
Auto-Submit: Michael Tatarski <mtatarski@google.com>
Cr-Commit-Position: refs/heads/main@{#1537226}