Add unit tests for post-migrate-targets.py code handling pyl data.
An upcoming change will switch the code handling pyl data to operate on
nodes from the pyl library. This moves that code to a separate module
with a unit test so that it can be updated with confidence that the code
continues to work.

Bug: 433525769
Change-Id: I99201bd5b115498cc46854f38f8a6127f03627cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7093422
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Haiyang Pan <hypan@google.com>
Commit-Queue: Haiyang Pan <hypan@google.com>
Cr-Commit-Position: refs/heads/main@{#1537408}