Mark the insets consumed by WebView shell as consumed
WebView shell applies padding to its content view based on the system
bar insets but was returning the same insets object passed to it without
marking the insets it had used as consumed. This means that when WebView
receives the same insets, it passes them to the web contents and they
get accounted for twice. This CL consumes the insets used by the content
view and enables DrawCutoutEdgeToEdge which was blocked on this bug.

Bug: 441759235,409721052
Change-Id: I0690f6d36418e96b9093c73b2319509d54fae74b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7091914
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Commit-Queue: Richard (Torne) Coles <torne@chromium.org>
Auto-Submit: Alex Mitra <alexmitra@chromium.org>
Commit-Queue: Alex Mitra <alexmitra@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537419}