Clarify AlphaType for Android hardware video decode
ComputeTextureTarget on Android always return TEXTURE_2D [1], so the
only way to create ClientSharedImage with TEXTURE_EXTERNAL is to use
this [2] ctor. There are 3 call-sites of that ctor, two of them are
windows specific and [3] is android.

It's hardware video decode images, they are always RGBA8888 [4] and
VideoFrames are marked as PIXEL_FORMAT_ABGR [5].

The reason for this is that hardware decoded videos use external
formats that are not expressed in SharedImageFormat at this point, but
images still require external sampling.

Make this more explicit, to clarify the intent.

[1] https://source.chromium.org/chromium/chromium/src/+/main:gpu/command_buffer/client/client_shared_image.cc;drc=b59fda5239d7637fdf7adbb2ea5b526aaf2cb7b4;l=92
[2] https://source.chromium.org/chromium/chromium/src/+/main:gpu/command_buffer/client/client_shared_image.h;drc=8900162810d457b64286ce653cb61bc2009c5068;l=349
[3] https://source.chromium.org/chromium/chromium/src/+/main:gpu/ipc/service/gpu_channel_shared_image_interface.cc;drc=8900162810d457b64286ce653cb61bc2009c5068;l=99
[4] https://source.chromium.org/chromium/chromium/src/+/main:gpu/command_buffer/service/shared_image/android_video_image_backing.cc;drc=8900162810d457b64286ce653cb61bc2009c5068;l=35
[5] https://source.chromium.org/chromium/chromium/src/+/main:media/gpu/android/video_frame_factory_impl.cc;drc=8900162810d457b64286ce653cb61bc2009c5068;l=155

Change-Id: I622041298f166c54b11bfd5056fc268174609858
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7079157
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537532}