Persist pending logs but don't close when flushing metrics
What we really want is to make sure if the process is killed we don't
lose logs - we don't actually need to close the current log when Chrome
is paused or going to the app switcher. Closing logs was causing a
large increase in the number of created and dropped logs.

Bug: 400975101
Change-Id: I5cc3f82d231b5b0203a5f5a956e29cbb6c0c70ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7091893
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Anand Ravi <anandrv@google.com>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537553}