IDB: compress rows in SQLite store
Implements a straightforward compression of IDB values prior to storing
on disk. See doc linked in crbug for discussion.

Prior testing has shown ZSTD to have performance (time, compression
ratio) comparable to or slightly better than Snappy. ZSTD is the
preferred compression algorithm for this reason and because its
performance is tunable, in addition to having some more sophisticated
features of potential future use. We will be able to perform real-world
experiments with compression level and some other parameters (and also
compare to Snappy), but for now some reasonable defaults are assumed.

Snappy compression is also included in this patch as a fallback for
size-constrained platforms (Android, Fuchsia), which do not already
include ZSTD's compression library, although they do include
decompression. (In this change, the decompression code is included on
those platforms because it's cheap/easy to include, and on the off
chance someone copies data from a Chromium profile on a system that did
use ZSTD for compression to one of these devices).

Since the SQLite backing store is not live, this shouldn't affect real
users. We will soon start deploying the SQLite store to real users for
in-memory profiles (such as Incognito), which is an ideal testing ground
for compression performance.

Test coverage is provided by existing tests, and there will be more
specific tests added when we add the ability to tweak and monitor
behavior (for experimentation).

Bug: 436887362

Change-Id: I7c0691b224ed766df6c6de50b860376e6d589dfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7080983
Commit-Queue: Evan Stade <evanstade@microsoft.com>
Reviewed-by: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537606}