[Contextual Tasks] Use <webview> instead of <iframe>
Since <webview> has some implementation differences with <iframes>,
switch back to using <webview> since that is the intended launch
product. These requests will be unauthenticated until the OAuth flow is
setup. To circumvent this, I added the login page as a valid URL to load
in the <webview>. This will allow devs to authenticate and get the
cookies if needed.

Bug: b:452978117
Change-Id: I2ae676915d632cfd32d6930f794f0262a44e160a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7095744
Commit-Queue: Duncan Mercer <mercerd@google.com>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537625}