[owners] Remove asully@chromium.org from ui/base/clipboard/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/asully@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I40c04a9b01daf20d9c8e6ef917aa61c95d9bd220
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7059781
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537626}