flyout: Let ListMenuHost own HierarchicalMenuController
Since `HierarchicalMenuController` is responsible for managing windows,
it was a little bit out of place to have it placed in
`BasicListMenu.java`. We now let `ListMenuHost` (who implements
`FlyoutHandler`) create and own this instance.

This refactor will allow us to do the refactor
where we let `FlyoutController` manage the list of windows.

There should be no functional change with this CL.

Bug: 433410990
Change-Id: I7689f09602a605d67050e82159505bc338f26d2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7093656
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Eleanor Lee <eleanorlee@google.com>
Commit-Queue: Masa Fujita <massan@google.com>
Cr-Commit-Position: refs/heads/main@{#1537728}