Map Chinese locales to language packs on ChromeOS
System locale strings like "zh-TW" are sometimes used as target
languages for SODA language packs, but Chinese language packs use a
specific naming convention encoded in MaybeMapToChineseLocale. This
mapping was being applied in non-ChromeOS SODA, but not in ChromeOS
SODA. This CL fixes this.

Added a test that captures this failure, and is fixed by this CL.

As part of this change, the MaybeMapToChineseLocale is updated to use
exact case for Chinese language locale strings to which it maps. This
has no effect on existing uses of this helper, which either only operate
on the first three characters (already lowercase) or else convert the
output to lower case already.

Bug: b:447469292
Change-Id: Idca99e670e6e80b31a054ec628bb35aeb1be6170
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7093778
Reviewed-by: Jennifer Ling <hsuanling@google.com>
Commit-Queue: Andrew Moylan <amoylan@google.com>
Reviewed-by: Evan Liu <evliu@google.com>
Cr-Commit-Position: refs/heads/main@{#1537732}