[//cc] Eliminate TileDisplayLayerImpl short-circuit on no tilings
TileDisplayLayerImpl::AppendQuads() currently short-circuits out if
`tilings_` is empty. However, PictureLayerImpl does not have any such
corresponding check. The check in TileDisplayLayerImpl comes from the
fact that a unittest crashes when trying to access the effect node for
the layer with there being no tilings. However, investigation shows
that that is just because we didn't properly set up the effect node
index in the test. In production, the effect node index will always be
synced over from the client. Hence, by hardening the unittest we can
eliminate the need for this explicit short-circuit.

Bug: 450651370
Change-Id: I4a79578ae989b80262df33bcb166dbf3cce875a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7095842
Reviewed-by: Victor Miura <vmiura@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537797}