[//gpu] Remove D3DImageBacking::CreateFromSwapChainBuffer()
This method is no longer used. This renders
D3DImageBacking::is_back_buffer_ redundant with
D3DImageBacking::swap_chain_ as the latter is set to non-null only in
the CreateFromSwapChainBuffers() method where the former is
unconditionally set to true, and the only usage of `is_back_buffer_` is
to check whether it's false in the case where `swap_chain_` is non-null.
This CL thus removes `is_back_buffer_` as well.

Bug: 415968760
Change-Id: I051147d85a3a05cfb0adf9cd4cd71197912e8dad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7094922
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537808}