[iOS] Update Favicon early return logic
This CL updates how the FaviconLoader returns a default value early.
Except for the AIM, no component was using the default image value.
This CL removes it to replace it with a Monogram, which is the default
state if no favicon has been fetched.

To fix the AIM mode, a default value is added to the implementation.

Bug: 443034511
Change-Id: Id39e9233f22671558e4a3a971445ad4795d29a3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7093178
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537813}