Trigger server metadata delete on deletion of server entities.
The CL addresses the issue of delete signals for AUTOFILL_VALUABLE
coming from sync not propagating to a corresponding delete request for
the AUTOFILL_VALUABLE_METADATA. Due to having different types for each
type of data, communication between the two bridges is necessary. In
this CL the ValuableSyncBridge is updated to trigger a server metadata
deletion notification when a server valuable is deleted.

Bug: 456156137, 436551488
Change-Id: I07110fc5e94b3f5a4b55699249fa3e285245369f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7096261
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Norge Vizcay <vizcay@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1537860}