segmentation_platform: base/hash/md5 -> crypto/obsolete/md5
There should be no change in behavior. The previous and new
implementations are equivalent.

Required for backward compatibility with existing on-disk data. Migrate
to the non-deprecated API.

This could be migrated to a safer hash in the future with a proper
migration strategy.

Bug: 406729261, 455854083
Change-Id: I3196de19adf47f5a8c8da64f4b3d48946a617102
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7091793
Commit-Queue: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Siddhartha S <ssid@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537933}