[Omnibox][WebUI] Add some infrastructure to enable switching to AI-Mode.
This allows for switching between the normal suggestions list and
AI-Mode.

Added a simple views::View container for the content of the popup
widget. Two WebViews (OmniboxPopupWebUIContent) are inserted into this
view and sized to match the container. To display the results list, that
view is made visible and the AI-Mode view is hidden. To show the
AI-Mode, the suggest list view is hidden and the other is made visible.
The selection of the view is plumbed through the OmniboxEditModel.

Current issues:
. A slight annoying flash on initial popup as WebUI is
initially resized.
. When AI-Mode is shown, escape or clicking away
doesn't close the popup.
. Pressing the Close(X) doesn't do anything.
.When AI-Mode is shown, resizing the browser window will revert back
to the results list.
. No theming visible in AI-Mode.

Change-Id: I2e548da961369721be82df31d710cca7127e96c8
Bug: 454080577
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7076151
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Allen Bauer <kylixrd@chromium.org>
Reviewed-by: David Yeung <dayeung@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1537963}