[IOS][Multiline] Fix multiline multiline toolbar broken height
When multiline is enabled the toolbar height is handled by the height
delegate.
This CL fixes an issue where the location bar shrinks when the user
backgrounds the app and opens it again.

This is caused because we are resetting the height on
updateLocationBarHeightForFullscreenProgress call even when the
location bar is focused.

Before: http://shortn/_bfCWsXd4ZZ
After: http://shortn/_4RGzlHHatW

Fixed: 454901242,454935540,454935736
Change-Id: Ibe8b2444db67ca8caf386c8be2aba0e2691c0398
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7098920
Auto-Submit: Ameur Hosni <ameurhosni@google.com>
Reviewed-by: Radu Nitescu <radunitescu@google.com>
Reviewed-by: Christian Xu <christianxu@chromium.org>
Commit-Queue: Ameur Hosni <ameurhosni@google.com>
Cr-Commit-Position: refs/heads/main@{#1537990}