[AIM] Focus composebox on follow-up region query.
This change ensures the composebox input is automatically focused when a
user performs a region search while already in an AIM session. This
allows users to immediately start typing follow-up questions about the
new region without needing to manually click the input field.

New Mojo plumbing was added to allow the browser process to request
focus on the composebox element in the WebUI.

Bug: 456231679
Change-Id: Idf959c9bc5a82f0775b00212b5c98dfd29f750d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7097958
Reviewed-by: Nihar Majmudar <niharm@google.com>
Commit-Queue: Juan Mojica <juanmojica@google.com>
Reviewed-by: Duncan Mercer <mercerd@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538013}