[a11y] Add freshness check for Java accessibility node cache
This involves an un-cached query to C++ on every cache hit, so will
negatively impact performance, but is meant to be a temporary and
small-scale experiment to evaluate the correctness of the cache
invalidation on the Java-side `AccessibilityNodeInfo` cache.

Histogram generation is gated behind the new
`AccessibilityCheckJavaNodeCacheFreshness` feature flag, and also
excludes the existing `AccessibilityDeprecateJavaNodeCache` experiment.

Bug: 427239087
Change-Id: I17c598c7f726ddf709290a67a78f767d4d488d65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7056065
Commit-Queue: Aaron Moss <aaronmoss@google.com>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Evan Liu <evliu@google.com>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538029}