[BrowserControls] Let browser take over offsets
When constraints / browser controls minHeight change when the layers are
hidden, we need to send the negative yOffsets to the hidden layers to
ensure they don't have one frame being displayed at their resting
position.

This CL fix this by always calculating the yOffset even when BCIV is
applicable, and only apply the resting yOffset when we are ready to
dispatch.

Bug: 417238089
Change-Id: I7b542e41245ab3e9c53cc448f4cc7b6a11eb49a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7097922
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Peilin Wang <peilinwang@google.com>
Cr-Commit-Position: refs/heads/main@{#1538036}