Revert "[TSLO] Update desktop windowing mode heuristic to account for min height"
This reverts commit 29352693fcdb1b37a96854a2535551cfaf1174c4.

Reason for revert: The new heuristic could be very restrictive; for example, on my Samsung tablet, the header height is 78px and the minimum tab strip height is 80px, so we fail to draw custom headers due to this min height threshold heuristic that I think is undesirable. Without this heuristic, the system UI controls / app UI might be slightly misaligned but this is more acceptable than not drawing the custom headers altogether.

Original change's description:
> [TSLO] Update desktop windowing mode heuristic to account for min height
> requirement
>
> OBSOLETE_HISTOGRAM[Android.DesktopWindowHeuristicResult4]=Replaced by Android.DesktopWindowHeuristicResult5
>
> Bug: 418264971
> Change-Id: Ie15c5c4add559125e2693bc721f81303fa791d79
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7023152
> Commit-Queue: Aishwarya Rajesh <aishwaryarj@google.com>
> Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1527683}

Bug: 418264971
Change-Id: I610c9fc60bd6b0a81c851cf843a3629b9f5077f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7101124
Commit-Queue: Aishwarya Rajesh <aishwaryarj@google.com>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538134}