[SidePanel] Update GetCurrentEntryId() usage to IsSidePanelEntryShowing()
A lot of existing usage of SidePaneUI::GetCurrentEntryId() is to
determine whether a type of side panel is currently showing. This CL
replaces the call to use SidePanelUI::IsSidePanelEntryShowing() instead
so that it is more clear that the callsite actually wants to know if a
given side panel is showing.

Bug: 445442616
Change-Id: I62e56c8cb2653d2415fff79a60577873c0b9d18f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7076304
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Commit-Queue: Steven Luong <stluong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538138}