[rustmojo] Add public api for the crate which caches wire formats
Followup to crrev.com/c/7068293. This CL implements the MojomParse
trait for a bunch of primitive types, and shows how we're going to use
it by adding two functions that use the trait.

These functions take (or return) values of types that implement the
trait, using its provided functions to get the mojom types and packed
format that the actual parsing/deparsing functions need.

Since we need to compute the packed format twice per message (every time
we parse or deparse!), we add a helper function to cache the results.

Bug: 444509367
Change-Id: Iba1e8a6b3bb1c6d6c908f698fed08c87a438f0ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7068394
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#1538179}