Preserve fullscreen state upon ADDED_TO_WORKSPACE event in tablet mode.
This is a regression in crrev.com/c/4520028, and happens to web
apps that request this event after creation.

This also create new browsertest that checks that the kiosk window is fullscreen on tablets.

This CL also removes the redundant parameter target_state of
AdjustStateForTabletMode

This browsertest runs through the kiosk launch logic and checks that a
device that started off in tablet mode ends up with a fullscreen browser
window.

Bug: 453589574
Test: KioskBlockTabletModeWebTest
Change-Id: I7fabdd0d6d0585f1f4f6b75c1c0677bc2681aa91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7092385
Reviewed-by: Ben Franz <bfranz@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538226}