Escape GuestViews for main frame detection in `GetAIPageContent()`
This change ensures that main frames within a GuestView are correctly identified as having a parent, treating them akin to subframes, with  different options to be applied. This will also be used by a later cl (https://crrev.com/c/7082911 ) to not block on slow main frames if they're in a GuestView.

Change-Id: I98b1b27f58b97f6e489a3a0468c5788410deabd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7101405
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538229}