Refactor `StreamModel::Update` method to avoid duplicate map lookups.
The previous implementation used `contains()` followed by `operator[]`,
which could result in two lookups for the same key. This is replaced
with `try_emplace`, which performs a single lookup and constructs the
element in-place only if the key doesn't already exist.

This is a minor optimization with no change in behavior.

Bug: 452950815
Change-Id: I7a27d7db398bc44b07b6ccd5ee97848505526fc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7093272
Commit-Queue: Jeffrey Yu <yuje@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538241}