[Bookmarks Bar] Update scene layer offset when coming on screen
This CL updates the scene layer offsets for the bookmark bar when the
bar is being animated back on screen. In some cases, BCIV is not able
to use the OffsetTag to set the correct y-position. In these cases, the
view/client must set them manually. One case is the top controls being
scrolled offscreen, and a link being pressed. The link should open a
new page and the top controls should scroll back on screen. When in
this state, in order to scroll with the rest of the top controls, the
bookmark bar needs to manually update its offset.

Bug: 455019009
Change-Id: Ia2246ec32d9e7f29131d98cd5ba0d0ba8c8649d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7103489
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538344}