[chromecast] Refactor how LocalService stores Binder mappings.
The previous version used static HashMaps from Intents to Binders
and from Binders to objects, which could be leaky.

The new version uses the fact that Android is already internally
storing mappings like these. The LocalService base class now
wraps whatever object it gets from bind() in a Binder instance
that it internally downcasts in connect(), leaving the unsafe
casting as an internal detail that doesn't leak to the
implementors and clients of Services that extend LocalService.

Bug: 301351406
Test: cast_base_junit_tests
Change-Id: I20d2b26198591b41be1730149f1c945ef65bc526
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7038661
Commit-Queue: Simeon Anfinrud <sanfin@chromium.org>
Reviewed-by: Sandeep Vijayasekar <sandv@google.com>
Cr-Commit-Position: refs/heads/main@{#1538399}