[chromecast] Make Observable.not()'s return type explicit.
The ? return type is ok in most use cases, but breaks when taking
`not(foo).or(not(bar))`, where `or()` requires the `T` parameter
of both inputs to be the same. When `not()` returns an
`Observable<?>`, the compiler is not allowed to see that the data
types actually are the same.

Bug: 301351406
Test: cast_base_junit_tests, cast_shell_junit_tests
Change-Id: I957eeb3d79ad4116652661545cd55fab0ec65898
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7038664
Reviewed-by: Sandeep Vijayasekar <sandv@google.com>
Commit-Queue: Simeon Anfinrud <sanfin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538405}