[chromecast] Add Observable.any().
This is a shortcut for not(not()), and is useful for situations
where an observer needs to be notified the first time some data
is activated and the time the last data is deactivated. Think of
it as a kind of reference-counting, and applicable to situations
where there may be multiple reasons to activate an observer, but
the observer should only be activated once at a time. E.g.

```java
Observable<?> requestKeepScreenOn();
Observable<?> audioPlayingWhileBackgroundModeNotSupported();

Observable<?> shouldKeepScreenOn() {
  return any(
          requestKeepScreenOn()
              .or(audioPlayingWhileBackgroundModeNotSupported()));
}
```

Bug: 429081455
Test: cast_base_junit_tests
Change-Id: I84dd520f2e155300b4f4ac675f3b0717c203eb7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7038666
Commit-Queue: Simeon Anfinrud <sanfin@chromium.org>
Reviewed-by: Sandeep Vijayasekar <sandv@google.com>
Cr-Commit-Position: refs/heads/main@{#1538408}