[//ui] Port WaylandSurfaceFactory test's call to CreateNativePixmap()
Move it from passing BufferFormat to passing SharedImageFormat. NOTE:
This necessitates making the calls on the object as a
SurfaceFactoryOzone*, as otherwise tries to match the calls against the
(virtual and overridden) variant taking in BufferFormat. This problem
will go away when we eliminate the latter (and make the former virtual).

Bug: 450325664
Change-Id: Id505855c246ddb4fa97ef7400b5abf6d8de0b487
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7099026
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538442}