[WaaP] Fix crash due to duplicated binding of PageMetricsHost
https://crrev.com/c/7047465 introduces a global binding for the mojom
interface PageMetricsHost that applies to all WebUI.
But it doesn't remove the existing binding from the experimental
WebUI version of ReloadButton, which is gated behind a flag
`kWebUIReloadButton`.

This CL removes the individual registry of the mojom interface
`PageMetricsHost` from `ReloadButtonUI`, removes the `BindInterface()`
from it, and directly supplies a MetricsReporter pointer to
ReloadButtonPageHandler by retrieving `MetricsReporterService` from
WebContents, which uses the global binding of PageMetricsHost.

Bug: 448794588
Change-Id: I55f1b2847365099fb48d363ab895a226c0b07b00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7105238
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Ming-Ying Chung <mych@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538452}