legion: Add SecureSessionImpl for Noise NN handshake and secure messaging
This change introduces `SecureSessionImpl`, which implements the
`SecureSession` interface. It uses the Noise protocol (NN handshake
pattern) to establish a secure session and then uses a `Crypter` for
encrypting and decrypting messages. Unit tests are added to verify the
handshake and secure communication flow.

Bug: b:444610593
Change-Id: Ib52d8960f0edd53b127a107896f4d385d22de32c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7071518
Commit-Queue: Oleh Lamzin <lamzin@google.com>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Cr-Commit-Position: refs/heads/main@{#1538468}