[iOS] Check FET Criteria For Gemini Contextual Chip
Uses the FET configuration for Gemini Contextual Chip to ensure that the
chip should be allowed to show. Also checks BWGService to ensure a page
is eligible for Gemini usage and checks PrefService to ensure a user has
completed the Gemini FRE promo.

Fixed: 455601523, 453980932
Change-Id: I58ac88a372a98dc43181370a67d1af69722eeebf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7092072
Commit-Queue: Joemer Ramos <joemerramos@google.com>
Reviewed-by: Adam Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/main@{#1538555}