Fix re-creating the History table after a raze
The CL https://crrev.com/c/6945226 added a call to GetVersionNumber in
RazeDbIfTooOld, called at the beginning of HistoryDatabase::Init.
It turns out that if the database has no meta table, calling
GetVersionNumber invokes the DatabaseErrorCallback, which razes the
database. This can happen if the database was previously razed, for
instance if the database was found to be corrupted. Because
DatabaseErrorCallback schedules the raze via an asynchronous task,
the raze will happen after Init completes. As a result, the table never
recovers from the first raze and get razed again.

This fix avoids calling GetVersionNumber of there is no meta table. This
allows Init to proceed without triggering the error callback. This is
how MetaTable::RazeIfIncompatible [1] is implemented.

[1]: https://crsrc.org/c/sql/meta_table.cc;l=113-115;drc=c6742c548ec7e9cbf20db834b966bfddc8fac909

Change-Id: I1a1e46edd5af0aa44e824837ad57a6586b763062
Bug: 369635654
Fixed: 456675138
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7092531
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538585}