Add flag to bypass beforeunload dialog when Actor is active
This change introduces a mechanism to allow the Actor to navigate a page
even if a `beforeunload` handler is present, without showing the
confirmation dialog to the user.

A new feature flag, `kGlicSkipBeforeUnloadDialogAndNavigate`, is added,
disabled by default.

When this flag is enabled, the `ChromeContentBrowserClient` will
instruct the `RenderFrameHostImpl` to bypass the `beforeunload` dialog
if the Actor is currently active on the associated WebContents. This is
achieved by implementing the new
`ContentBrowserClient::ShouldSkipBeforeUnloadDialog` method, which is
called from `RenderFrameHostImpl::RunBeforeUnloadConfirm`.

This prevents the Actor from getting stuck on pages that use
`beforeunload` to prompt the user before leaving.

Browser tests and unit tests are included to verify the new behavior.

Change-Id: Ief44ea56b2d36df73e1587bf5c1b5aae77b8172d
Bug: 439600688, b: 454961534
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7080464
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Auto-Submit: Aaron Selya <selya@google.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538742}