Avoid restoring subframes pending deletion from BFCache
When DelayRfhDestructionsOnUnloadAndDetach is enabled or when unload
handlers are delayed, it is possible that some subframes are pending
deletion at the time when the main frame is considered for the BFCache.
The main frame can be admitted, but the subframes should never be
restored.

Bug: 327446163
Cq-Include-Trybots: luci.chromium.try:linux-oi-rel,linux-bfcache-rel
Change-Id: I889ba1a2c341aaccd83ea8e12de03a55bcf0699f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7028393
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Aldo Culquicondor <acondor@chromium.org>
Auto-Submit: Aldo Culquicondor <acondor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538829}