[ntp-next] Listen for change of the Chips display pref
This CL adds the pref listeners for the NTP handlers and methods to
change the visibility of the action chips in the web ui on pref toggles.

It also adds an additional fail-closed condition to the rendering of the
NTP action chips which hinges on whether AIMode is enabled for the
profile.

Change was tested locally, where toggling the "Show Chips" option allows
the Action Chips to be displayed/hidden in the DOM:
https://screencast.googleplex.com/cast/NDY5NDg5MDM5ODA4OTIxNnw4NjljNmExOC1jNw.

Bug: 450291144

Change-Id: If1509a8be79d0dbdc9bd9b96f4617052453a0aea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7088588
Commit-Queue: Gianni Chen <giannich@chromium.org>
Reviewed-by: Riley Tatum <rtatum@google.com>
Reviewed-by: Jennifer Serrano <jennserrano@google.com>
Cr-Commit-Position: refs/heads/main@{#1538830}