[PWA] Add timeout for dev-specified page manifest fetch
The PageManifestManager::GetSpecifiedManifest call will wait until a
manifest url is specified, so if no url is specified it will wait
forever. To prevent timeouts for commands, this change adds a timeout to
the WebAppDataRetriever, which wraps that call. We wait 30 seconds here.

This values seems reasonable as this is the value used for the
WebAppUrlLoader too.

R=mek@chromium.org

Fixed: 456534052
Change-Id: I75fe6ba5bea54994496132f8c02590f02e1f9cd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7103642
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Auto-Submit: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538851}