Pass FrameTreeNodeId instead of WebContents to ContextualTaskUiService::HandleNavigation()
The HandleNavigation call will post a task to process the navigation. As a result, the WebContents may get deleted when the posted task runs.
This CL changes the call to take a FrameTreeNodeId instead of
WebContents, so the posted tasks can check whether the WebContents
is still around

Bug: 449161572
Change-Id: I74f2f95f94163177852a513ef139a779a85f51dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7106293
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538888}