[Settings Search] Send Query to SettingsIndexData to perform Search
This change adds a registry of the available prefs (for now) and uses
them to index the xml data. Currently, the logic is too naive and is
meant to be built on in next steps so it is able to determine what prefs
are visible to the user and can be retrieved when querying. This is fine
for now.

Bug: 444470792
Change-Id: I5587803b959e3d3fbe4ad7a250347d2b47fffeef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7103722
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Moe Adel <adelm@google.com>
Cr-Commit-Position: refs/heads/main@{#1538895}