Delay SW loader deletion until fetch clone is complete
When using `race-network-and-fetch-handler`, the
ServiceWorkerMainResourceLoader could be destroyed after the network
response is received, but before the fetch handler has finished cloning
the response body. This could lead to a race condition where the loader
is deleted prematurely.

This change ensures that the loader's lifetime is extended until the
response body has been fully cloned for the fetch handler or the mojo
disconnection to the ServiceWorker. A callback is introduced to signal
the completion of the cloning process or disconnection, and the deletion
logic is updated to wait for this signal before invalidating the loader.

This behavior is gated behind the
`kServiceWorkerStaticRouterRaceRequestFix2` feature flag.

Change-Id: If7ec62b18c99c8fc13faf657a44969a42bf7f82a
Bug: 340949948
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7079632
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Keita Suzuki <suzukikeita@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1538932}