Adds a new Glic promo bubble that opens Glic.
It replaces the existing snooze type IPH with a custom action IPH, when
the new feature IPH_GlicTryIt is enabled. If IPH_GlicTryIt and
IPH_GlicPromo are both enabled, only IPH_GlicTryIt will be triggered by
GlicIphController.

The GlicTryIt promo does not attempt to prewarm the FRE as it would be
accepted too late for prewarming to have an effect (the FRE would be
opened immediately upon selecting the "Try It" button.

N.B. the strings are marked as translateable=false to be able to merge
into release branches for use in en-US only.

Bug: b:454319747
Change-Id: Ibafb3f92eb4d336cb34b310b89c7de702cdc0bbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7075832
Commit-Queue: Justin DeWitt <dewittj@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539089}