[//cc] Move PictureLayerImpl resourceless SW draws into helper method
Paves the way for moving detection of this case into
TileBasedLayerImpl::AppendQuads() in a followup and having that method
call into the subclass to handle it, thus removing this case from
PictureLayerImpl::AppendQuadsSpecialization() altogether. This will
help in deduping the flow of that method with that of
TileDisplayLayerImpl's version.

Bug: 450651370
Change-Id: If9e0655383457a7a29e5a41464d1b2a405f65309
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7106318
Reviewed-by: Victor Miura <vmiura@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539220}