policy: reduce retry frequency for FM token upload
Increase the delay for the first retry from one to five minutes.

Also, the backup logic was not actually verified in test, i.e. it did
not check if the retry was fired earlier than expected (which it did).

Thus, this CL also adjusts the retry logic and the test.

Bug: b:455916498
Change-Id: If1f52bfe2b66fa565c90ed35f599dd787f1cba59
Tests: Adjusted unit test
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7105760
Commit-Queue: Roland Bock <rbock@google.com>
Reviewed-by: Artem Sumaneev <asumaneev@google.com>
Cr-Commit-Position: refs/heads/main@{#1539256}