[dbsc] Add a method to copy keys between services
This change introduces a new method `CopyKeyFromOtherService` to the
`UnexportableKeyService` interface. This allows for an unexportable key
to be duplicated from one service instance to another.

The implementation retrieves the wrapped key data from the source
service and uses it to construct a new key in the destination service.
While the underlying cryptographic key material is identical, the new
key will have a distinct `UnexportableKeyId`.

Bug: 455537760
Change-Id: I2fe762e46da022ee27c7943cfc1468ba84759e82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7106138
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Ernest Nguyen Hung <ernn@google.com>
Cr-Commit-Position: refs/heads/main@{#1539266}