Refactor CodeCache StoragePartitionImpl tests
Rewrite RemoveCodeCacheTester so that it uses a CodeCacheHost to
add/find entries in the code cache rather than directly using
implementation details of GeneratedCodeCache. This allows the tests to
operate regardless of how CodeCacheHost is implemented. This fixes test
failures when the UsePersistentCacheForCodeCache feature is enabled, as
GeneratedCodeCache is not used in that configuration.

Fixed: 450397707
Change-Id: Id3e8824aa97433a625ca40a8de344aba5eba148c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7105738
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Greg Thompson <grt@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Auto-Submit: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539295}