[ios] Fix unsafe buffer in IOSChromeVariationsSeedStore
Remove the static array and instead use an helper function used to
compute the next SeedApplicationStage for a given value (constexpr).

Mark IncrementSeedApplicationStageIfAt(...) which allow the compiler
to elimitate the std::optional<...> and the CHECK when optimising.

This allow the following call:

  IncrementSeedApplicationStageIfAt(SeedApplicationStage::kNoSeed);

to result in the same code as if written like this:

  if (gCurrentStage == SeedApplicationStage::kNoSeed) {
    gCurrentStage = SeedApplicationStage::kSeedStored;
  }

Bug: 40285824
Change-Id: Idae3ce88331a7b71225d8ca02bfb1fa4d9b12f19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7101379
Reviewed-by: Ginny Huang <ginnyhuang@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539304}