[Unowned User Data] Migrate AppBrowserController
This migrates AppBrowserController to Unowned User Data, so it can be
accessed from a (forward-declared, even) BrowserWindowInterface rather
than needing to be retrieved from a Browser or BrowserView, and so it
can be mocked more easily.

The number of includes in app_browser_controller.h is also reduced,
specifically to avoid bringing in all of browser_window_interface.h,
in order to minimize downstream impact on files that now need to
include it.

This does not remove all of the existing accessors, though it does cut
down on the number of such accessors, in an effort to reduce the number
of ways one could retrieve the object. Future CLs can further cut down
those APIs to eventually use only methods on AppBrowserController
itself.

Change-Id: I0fb96cecba5ed8811e02516e94594da177da5579
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7045837
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Tom Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539330}