Use internal popup menus for tabs in actor-controlled states
HTML <select> element dropdown menus are missing from
RenderWidgetHostView::CopyFromSurface() captures on Mac, since they are
rendered using native UI elements.  This change disables of the use of
native UI elements for dropdowns on Mac during actor-controlled states.
Chrome's internal rendering of dropdowns is tested on all platforms,
including Mac, in ExternalPopupMenu*Tests.

This change includes a browsertest to verify that dropdowns are visible
to CopyFromSurface(), and another parametrized browser test to ensure
that internal / external popup state correctly propagates to existing
and newly-created out-of-process iframes.

As part of this, fix a bug (crbug.com/456537756) in
RenderWidgetHostView::CopyFromSurface() where the operation never
completes if performed at the same time as dismissing an open popup --
this is fixed by this CL's changes to render_widget_host_view_base.cc,
which add missing calls to the completion callback. This fix is
necessary for the browser tests to not timeout.

Co-authored-by: Paul Jensen <pauljensen@chromium.org>
Bug: b:447164093, 454961534
Fixed: 456537756
Change-Id: I57960b1a7ef639654b8283d673c29b095a0303ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7074751
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Caleb Raitto <caraitto@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539427}