[Masonry] Fix masonry-grid-template-columns-computed-withcontent.html
This test was failing on three different checks. Two of which, the
actual output matched that of grid and were expected, so update the
expectations for those two.

There was also a check that setting 'none' for grid-template-rows on a
column masonry returned back a computed value of 'none'. This is not
expected, since there are no template rows in a column masonry. The
expected computed value is '', however, updating the expectation to ''
fails because an empty string gets overridden to the computed value in
the harness. Instead, just remove this entry since the test is for
grid-template-columns, not rows.

Bug: 343257585
Change-Id: I33211ed9e217f4b8605a370effac4782237dc8c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7114684
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Auto-Submit: Alison Maher <almaher@microsoft.com>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1539684}