[Contextual Tasks] Address crash when opening links from WebUI
The browser was crashing when clicking a link from a thread because
by the time the posted task ran, the WebContents was out of scope.
This patch addresses the issue by passing a TabInterface if it is
available instead (a longer-lived primitive). A few other updates
have been made to simplify some parts of the code including the
service holding a profile rather than querying for one.

Bug: 456772056
Change-Id: Id79f790dfbb6bd6f562e29148f4a6f7ccdf1ed3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7106437
Commit-Queue: David Pennington <dpenning@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539686}