[fedcm] Introduce a Navigation Throttle to allow IdPs to initiate FedCM
In this Cl, we introduce a subclass of Navigation Throttle and hook it
into the registry behind a chrome://flag.

The throttle allows IdPs to cancel a navigation to one of its
endpoints and initiating a FedCM flow without requiring
relying parties to change, by returning a special HTTP header to the
browser.

The implementation of the throttle is left to a series of subsequent
CLs and described in the design doc linked below.

Design Doc: https://docs.google.com/document/d/1Capfl_r7RKYixzxgiqoNXI394Z7zSirrECzXBR6GMlM/edit?tab=t.0

Bug: 455614294
Change-Id: Ie51fa3d12fbac304a8b32e464ef6bf6b813c21e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7088077
Reviewed-by: Nicolás Peña <npm@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Sam Goto <goto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539703}