[indexeddb] Report high number of PendingConnection/Connection to histograms.
This CL adds histograms with the number of
content::indexed_db::(Pending)Connection objects alive in the browser
process when a new (Pending)Connection is created. They are recorded
every time a (Pending)Connection object is created and the number of
existing (PendingConnection) objects is greater than 10k.

The histograms will be used to diagnose crbug.com/381086791 by:
- Making it possible to compute the proportion of clients with an elevated number of active (Pending)Connection objects.
- Acting as a trace trigger to understand how clients get into that state.

Due to the way the histograms are reported (on each connection creation
when the active number of connections is already high, rather than on a
heartbeat), the percentiles don't have a clear meaning. This doesn't
prevent us from using the histograms to achieve the 2 goals described
above.

Motivating data:
IndexedDB consumes 22% of allocated memory in browser processes
with >1GB of allocated memory. Most of this allocated memory is
directly related to connections or pending connections.
https://pprof.corp.google.com/?id=c335713120434b049e183792f2ee8dde&pivot=indexed_db&tab=flame

Bug: 381086791
Change-Id: I67b307a7adf96b36162e99e373f5495810c50c3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7101926
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539787}