Throttle the navigation request for the Android spare renderer
The CL adds a navigation throttle to defer the navigation request
until the priority of the Android spare renderer has been correctly
updated. The priority update of a process on Android is asynchronous.
Before the CL, the priority of the spare renderer can be kept at WAIVED
level when it is used for the navigation. This will cause potential
OOM kills and lead to a sad frame.
With the CL, the renderer will be ensured to be of VISIBLE priority when
the navigation starts. If it gets killed before that, the navigation
stack will create a new renderer after receiving the network response.

Bug: 447645527
Change-Id: I359767f8d272597995b5b51ea8c00182259b30e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7051342
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Jiacheng Guo <gjc@google.com>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Shin Kawamura <kawasin@google.com>
Cr-Commit-Position: refs/heads/main@{#1539798}