[//gpu] Remove GMBSupport dep from mappable_buffer_test_template.h
We now pass exactly the formats/usages that are supported on all
platforms except Ozone, where the underlying check performed by
GpuMemoryBufferSupport::IsNativeGpuMemoryBufferConfigurationSupported()
is still needed. Hence, we can replace usage of that method (which we're
trying to remove) with a new method internal to this test template that
does what's necessary here.

Bug: 404905709
Change-Id: I9a8af56fa34f07cdca9b98b1ca2482cf070cec21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7110907
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539841}