Gracefully handle broken "pending substitution" values
CL:7096301 tried to guard against invalid adjustments to the computed
value of writing-mode/direction by remembering the expansion-time value
in StyleCascade. That approach didn't work since the animations
machinery reads the computed value of writing-mode/direction directly,
which means we would need to do a similar trick there.

This CL reverts CL:7096301 (except the test and some comments),
and instead handles the previous NOTREACHED() branch as 'unset'.
This means we'll have functional bugs when we hit this path,
which is assumed to be acceptable until we can address the invalid
adjustment problem properly.

Bug: 40527196, 444818221
Change-Id: I3b28d435e650885ec82428f60be554345a7740c9
Fixed: 456502372
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7108170
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539875}