[RouteMatching] Full route-condition parsing support.
This adds support for and/or/not expressions in @route. See
https://wicg.github.io/declarative-partial-updates/css-route-matching/#at-route

Move @route parsing code into a separate file, and hook it up with the
common conditional parsing machinery (to build a tree of
ConditionalExpNode objects). Similarly, for evaluation, use the common
conditional evaluation machinery, so that expressions such as "not",
"and" and "or" (and parentheses) work.

Remove a few expectations from invalid-route-rule.html, and add better
tests. Rules easily parse successfully, since unrecognized content is
generally consumed by "<general-enclosed>".

Bug: 436805487
Change-Id: I8fb974ce5e905d36654355cc83eff8b787825659
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7102486
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539902}