Support activity recreation for multi-column mode.
In activity recreation case, fragments are recreated in the library side.
So, instead of making the instance, SettingsActivity finds the
MultiColumnSettings fragment from the fragment manager, and keep it
from its member.

Also, this CL restores detailed titles from the recreated fragments.
In addition, SingleCategorySettings lazily sets its title on resuming.
However, it is too late, and actually if the fragment is in the
back stack, it will not be resumed so title was kept unset.
This CL also moves it to done in creation.

BUG=455886870
TEST=Tryjob. Run on DUT with changing the system theme, which trigger
     activity recreation.

Change-Id: I89cda80cdf4c4e37d5dc250991809db584268a31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7100164
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539939}