[Autofill] Do not clear field values in TestBAM::AddSeenForms()
By default -- and we only use the default -- TestBAM::AddSeenForms()
clears the values of FormData before putting it into AutofillManager's
form cache.

This probably is a leftover from ancient times when
RetrieveFromCache() cleared AutofillField::value if they were
identical to the cached field's value.

This CL removes this implicit clearing of values in
TestBAM::AddSeenForms() and instead explicitly adds it to those tests
that fail without it.

Bug: 40183094
Change-Id: I99d769ef1bb50f86f7e062c54a20b83585a498ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7117820
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1539948}