Simplify pref handling in CreditCardSaveManager test
This is a "reland" of a change that was originally made as part of
crrev.com/c/7105961. The difference is that this change keeps the
DestroyAutofillClient() line, which is required to ensure that the
test sync service outlives the AutofillClient and the objects it owns.

Bug: 40100455
Change-Id: I317bd7a4bc8e2e529c1a1506913c332e56e5d3f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7117384
Reviewed-by: Christoph Schwering <schwering@google.com>
Auto-Submit: Jan Keitel <jkeitel@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1539955}