Parse XML documents through Rust XML crate
Introduce an XMLDocumentParserRs, which is instantiated in Document
depending on flag. The change in this location enables passing XML
parsing through the Rust XML crate, instead of using libxml2.

The implementation is based on logic and state behavior in
XMLDocumentParser - and behaves towards Document and the DocumentParser
parent class ideally in the same way.

This is by now fairly complete, reliably passing a wide set
of SVG, XMLHttprequest and XHTML tests, but there are a set of known
issues expressed as TODO's in the code, and tracked in the issue below:
* External entities need to be deprecated.
* XML Rust crate always sends a DocumentStart even if there is no <?xml PI.
* Entity parsing needs redeclaration of all XHTML and HTML5 entities

For now, keep the flag disabled and run the most important XML related
tests in a virtual test suite called rust-xml.

Adapt some unit tests to pass when the flag is on, partially due to
different error message wording between libxml2 and Rust XML crate.

More details in: http://go/chrome_x_mitigation (internal)

Bug: 441911594
Binary-Size: Binary size increase is expected to recover when completing transition and removing libxml2.
Fuchsia-Binary-Size: Binary size increase is expected to recover when completing transition and removing libxml2.
Change-Id: If794513791d91cbfa06927dad38cbdf90b7935b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7088396
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1539983}