[Startup] Use ApplicationStartInfo for Clank Startup Metrics
Starting with Android V, we now have ApplicationStartInfo that gives us
various bits of information regarding startup. One of the timestamps
provided is TimeToFirstFrame.

Local testing has shown that this metric, on cold starts, can be very
different from current cold start metric times. So this CL starts
recording the metric to see how stark the differences are.

Currently this is recorded regardless of NTP, web etc. startup as the
idea behind it is to give a general overview of user-perceived cold
start time. Future iterations can drill down a bit more if needed.

Cold start detection uses ApplicationStartInfo as well instead of our
current solution (ColdStartTracker). It unfortunately doesn't provide a
way of discarding starts that were backgrounded in between Activity
creation and this metric being recorded so
SimpleStartupForegroundSessionDetector is used for that.

Bug: 337886912
Change-Id: Ib226b2bbdf9e53630311a405a7d4ea8b11fde278
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7097681
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Sean Maher <spvm@chromium.org>
Commit-Queue: Nafis Abedin <nafisabedin@google.com>
Cr-Commit-Position: refs/heads/main@{#1540101}