[iOS] Remove memory cycle between Save to Drive coordinator and mediator
At the moment the mediator holds strong references to all command
handlers including the `_manageStorageAlertHandler` which happens to be
the coordinator. This CL makes these references weak as they should be.
The reference for this handler is now also set to nil in disconnect.

Change-Id: I93f4a980180ac891fd2ba5325cc7c5efe87ba8f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7117326
Commit-Queue: Quentin Pubert <qpubert@google.com>
Reviewed-by: Elmehdi Rahmaoui <erahmaoui@google.com>
Auto-Submit: Quentin Pubert <qpubert@google.com>
Cr-Commit-Position: refs/heads/main@{#1540114}