android-a11y: add focusable main test
This change demonstrates the tree built for a focusable <main> node. In particular, Chrome does not populate `text`, `content description` or alike for the main node itself. TalkBack is therefore responsible for the output generated in the bug attached and that bug needs to be routed thusly.

Bug: 364466909

Change-Id: I7183c893415a86b5acd29925dabb28ed8e8800f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7107342
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540169}