Let MailboxTextureBacking store ClientSharedImage instead of Mailbox
This CL refactors MailboxTextureBacking so that it now stores a
scoped_refptr<ClientSharedImage> instead of a Mailbox. This change is
to ensure the ClientSharedImage's lifespan is at least as long as
the MailboxTextureBacking, so that the ScopedRasterAccess that refers
to the ClientSharedImage is always valid.

Bug: 40286368
Change-Id: I1b030aff1059c2faae90879a261b8ab8ecffdadb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7096658
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Mingjing Zhang <mjzhang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540170}