[HiC] Fix cross-origin iframes with canvas content
Cross origin iframes that contain a canvas should be
able to draw content from that iframes origin but not the
main frame origin. Fix a problem where we were always using
the top-level origin as the one to compare to.

Add tests for iframes, background image and html images. The latter
confirm that image origin tracking works as expected.

Bug: 455905498
Change-Id: I81f53df00c40ee43a80a530d4a22eb157a821871
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7103440
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540203}