[GlicMI] Enable Multi-Instance flag for Actor tests.
Tests were failing with the multi-instance flag enabled because there
was no support for daisy chaining the instance when actor creates a new
tab, and that they test UI visibility which can be fragile in
Multi-Instance.

Another finding with the tests were that the GlicContentsElementId gets
torn down and replaced with an inactive embedder when the tab is
backgrounded, and the GlicContentsElementId doesn't switch between
instances when a tab is closed/foregrounded, so we should instead update
these tests to depend on the instance behavior rather than the UI.

A couple test suites keep Multi-Instance disabled - those are tracked
inside the bug.

Bug: 454665367
Change-Id: I776e34e3db401d702973e9613b841d9cb9057ace
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7091356
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Sana Akbani <sanaakbani@google.com>
Cr-Commit-Position: refs/heads/main@{#1540228}