Extend Virtualization UMA histograms
We would like to keep collecting the data for the histograms in this
collection. At the same time, we also deprecate collection for the
PLUGIN_VM variant as it is obsolete. Histograms other than
{Start,Stop}Duration are in fact not collected for variants other than
ARCVM, but there is no reason to remove that at this time.

Bug: b:451609797, b:451610802
Change-Id: I1a3658146767a583bc1e1620bca3bc916d9a670a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7068655
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Junichi Uekawa <uekawa@chromium.org>
Auto-Submit: maciek swiech <drmasquatch@google.com>
Cr-Commit-Position: refs/heads/main@{#1540258}