[InputVizard] Add histogram to track event time down time delta.
With this change we are adding a new histogram
`Android.InputOnViz.Viz.DroppedSequences.EventAndDownTimeDelta`, to
divide dropped sequences based on event time and down time delta. The
hypothesis that we want to verify is: some sequences are directly making
it to Viz sometimes without a transfer being requested for by the
Browser process. We expect the sequences that directly come on Viz from
system to majorly have 0 delta, modulo some event timestamps bugs in
Android itself.

Bug: 448836551
Change-Id: I7259819ba66dffdc4368fed5c9112c063107bc62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7117502
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Kartar Singh <kartarsingh@google.com>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540260}