[agents][eval] Store system prompt in `//GEMINI.md`
Currently, the system and user prompts are concatenated in that order
and fed to gemini-cli's stdin as one query. This breaks testing of slash
commands with system context because the command parser won't find the
leading `/` in the query that identifies the command [0].

Allow test cases to test slash commands by writing the system prompt to
the fake per-test `//GEMINI.md` [1], which gemini-cli will discover and
handle separately from the user prompt.

Reversing the prompt order won't work either because the parser will
treat the system prompt as one of the command `{{args}}` [2].

[0]: https://github.com/google-gemini/gemini-cli/blob/460c3debf5ec73f0652a496254ad9b5b3622caf7/packages/cli/src/nonInteractiveCliCommands.ts#L38
[1]: https://github.com/google-gemini/gemini-cli/blob/main/docs/cli/gemini-md.md#understand-the-context-hierarchy
[2]: https://github.com/google-gemini/gemini-cli/blob/460c3debf5ec73f0652a496254ad9b5b3622caf7/packages/cli/src/utils/commands.ts#L68

Bug: 454049467
Test: Check telemetry output file. Command is expanded with this change.
Change-Id: Iff1fb0b95f882dfc073b2dd1da6c937d289e20d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7120319
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1540313}