[owners] Remove btriebw@chromium.org from chrome/browser/window_management/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/btriebw@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: If467a5928abbf75e3490773bcbfd044f5b431ab7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7117356
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Mike Wasserman <msw@chromium.org>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540342}