[Extensions] Handle channel open failure when lazy context fails to load
When opening a message channel to a lazy context (e.g. a service worker)
the context must be started before the channel can be established. If
this startup process fails, the corresponding callback is invoked with a
null context (see https://crrev.com/c/6937175).

Previously, this failure case was not correctly handled, leaving the
channel in a pending state. This also means the port opener was never
notified of the connection failure. This can happen if e.g. a service
worker fails to start, or if the extension is deactivated with pending
tasks.

This patch adds logic to gracefully handle the failure. The pending
channel is now correctly cleaned up, and an `OnDisconnect` event is
dispatched to signal that the connection failed.

Bug: 371011217
Change-Id: I806fa7dfe98c9c92c8723ec3f5a2f8658976de19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7046053
Commit-Queue: Andrea Orru <andreaorru@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540421}