Reuse request NavigationClient on RenderFrame <> RenderFrame swap
With RenderDocument, a new NavigationClient is created even
when doing same-SIG navigation because the RenderFrame changes. This
could've regressed navigation cancellation behavior that depended
on reusing the NavigationClient (allowing cancellations even when
the navigation commit is in flight), which was the reason
RendererCancellationThrottle was introduced. However, we're
suspecting RendererCancellationThrottle is causing latency
regressions as it adds an IPC hop to wait for the navigation
initiator JS task to finish.

This CL attempts to maintain the navigation cancellation without
using the throttle, by actually reusing the requester
NavigationClient to commit, and transferring the ownership to
the new RenderFrame when the commit arrives on the renderer side.

Bug: 40615943
Change-Id: I29accddc4e800127c15f2b823dacf164b9a1139d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7082082
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540436}