Enable touchpad swipe-to-navigate everywhere in clank window
On desktop chrome, horizontal swipes everywhere trigger gesture
navigation, while prior to this change clank only do that for swipes
starting close to the window edge.
To provide an user experience closer to other desktop platforms, allow
gesture nav to be initiated everywhere if the swipe is performed on a
touchpad. Swipes on touchscreen will remain edge-initiated only to keep
it consistent with Android system gesture nav.

This could be disabled in the accessibility settings to keep the
existing behavior edge, on by default on LFF devices
(crbug.com/364370338#comment45, comment47).
The setting string is taken from the UX proposal:
https://docs.google.com/presentation/d/1EFN2O1aaFVZ7CSYL4LdpHS-fq9qmPEj2Arp8hbZ3MEQ/edit?slide=id.g33d19dd72d1_0_25#slide=id.g33d19dd72d1_0_25

On the logic added for the accessibility setting control:
* the value of user preference is stored by WebPreferences in Chrome
  * the default value is set in PrefsTabHelper::RegisterProfilePrefs,
    false on phone and true everywhere else
  * user can toggle the value in accessibility settings, which reads and
    writes to that user preference value via a delegate
* On creation of OverscrollController, RWHVAndroid reads and propagate
  value of the user preference to OverscrollControllerAndroid, and then
  OverscrollRefresh.
* On web preference value change, the same RWHVAndroid function will be
  called to update the OverscrollRefresh property.
* OverscrollRefresh::OnOverscrolled uses its internal property to
  determine to correct behavior

Define RenderWidgetHostViewBase::SetTouchpadOverscrollHistoryNavigation
and override it in RenderWidgetHostViewAndroid.
We can't cast the RenderWidgetHostViewBase instance to
RenderWidgetHostViewAndroid because tests use TestRenderWidgetHostView.

      2-finger swipe on touchpad when feature is enabled in \
      accessibilty settings, and not otherwise: \
      https://screencast.googleplex.com/cast/NTI2NTA3NDI2NzYxOTMyOHxkZWNkN2VlZi1lYg

Bug: 364370338
Test: clank fwd/back gesture nav triggered anywhere in the window with \
Test: ui/android/overscroll_refresh_unittest.cc
Change-Id: I9f9c08eceb32909bd3b753d03917677d86216633
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6826643
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>
Auto-Submit: Grace Cham <hscham@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540445}