YourSavedInfo: Make entries list editability configurable
On leaf pages of the 'Your saved info' section, we need the ability to
disable the addition of new entries based on a dedicated preference.

This change introduces the `allowEditingPref` property to
`autofill_ai_entries_list.ts`. This property, when provided, is taken
into consideration to determine the final state of the 'Add' button.

Bug: 451349831
Change-Id: Iccac22def1fabc78189f4a8b9793c642cbb99a68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7112800
Commit-Queue: Stanislav Mikheyev <mikhe@google.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540522}