[CPE][Signal] Skip passkey update for the same username
This CL adds an early return for user details update, if the stored
credential username is matching the one passed by the Signal API.

The only thing that changes with this CL is that CPE will not be
notifying the browser about the need to run the credential migration.
Credential migration would do nothing anyway, but it still seems
unnecessary.

Tricky to test as nothing actually changes for the credential, but it
should be fine given the method is generally unit tested.

Bug: 432260316
Change-Id: Ia7f6a9866a04f7e1364ffa77ed471c651199d095
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7124000
Reviewed-by: Adem Derinel <derinel@google.com>
Commit-Queue: Rafał Godlewski <rgod@google.com>
Cr-Commit-Position: refs/heads/main@{#1540541}