AnchorEvaluatorImpl: const / mutable cleanup.
If member functions that need to update member variables are not const,
the member variables don't need to be mutable. No idea how we ended up
here.

Also rename GetAnchorRect(), and make it non-const. It used to be a
simple getter, but that's no longer the case with the CSSAnchorUpdate
feature. Now it even has side-effects.

Get rid of some duplicate code in EvaluateAnchorSize() and
EvaluateAnchor() "while I'm at it".

Change-Id: I9e6308050d4840b4c113fe9dd34de3fb88a01686
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7123059
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540661}